Crate google_dlp2

This documentation was generated from DLP crate version 1.0.9+20190629, where 20190629 is the exact revision of the dlp:v2 schema built by the mako code generator v1.0.9.

Everything else about the DLP v2 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

info types

list

organizations

deidentify templates create, deidentify templates delete, deidentify templates get, deidentify templates list, deidentify templates patch, inspect templates create, inspect templates delete, inspect templates get, inspect templates list, inspect templates patch, stored info types create, stored info types delete, stored info types get, stored info types list and stored info types patch

projects

content deidentify, content inspect, content reidentify, deidentify templates create, deidentify templates delete, deidentify templates get, deidentify templates list, deidentify templates patch, dlp jobs cancel, dlp jobs create, dlp jobs delete, dlp jobs get, dlp jobs list, image redact, inspect templates create, inspect templates delete, inspect templates get, inspect templates list, inspect templates patch, job triggers activate, job triggers create, job triggers delete, job triggers get, job triggers list, job triggers patch, locations content deidentify, stored info types create, stored info types delete, stored info types get, stored info types list and stored info types patch

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.projects().deidentify_templates_delete(...).doit()
let r = hub.projects().inspect_templates_delete(...).doit()
let r = hub.projects().job_triggers_delete(...).doit()
let r = hub.projects().dlp_jobs_cancel(...).doit()
let r = hub.projects().stored_info_types_delete(...).doit()
let r = hub.organizations().stored_info_types_delete(...).doit()
let r = hub.organizations().deidentify_templates_delete(...).doit()
let r = hub.projects().dlp_jobs_delete(...).doit()
let r = hub.organizations().inspect_templates_delete(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-dlp2 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_dlp2 as dlp2;
use dlp2::GooglePrivacyDlpV2CancelDlpJobRequest;
use dlp2::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use dlp2::DLP;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = DLP::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = GooglePrivacyDlpV2CancelDlpJobRequest::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.projects().dlp_jobs_cancel(req, "name")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Chunk

ContentRange
Implements the Content-Range header, for serialization only

DLP
Central instance to access all DLP related resource activities

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DummyNetworkStream

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

GooglePrivacyDlpV2InspectJobConfig
There is no detailed description.

GooglePrivacyDlpV2CreateDlpJobRequest
Request message for CreateDlpJobRequest. Used to initiate long running jobs such as calculating risk metrics or inspecting Google Cloud Storage.

GooglePrivacyDlpV2RedactImageRequest
Request to search for potentially sensitive info in an image and redact it by covering it with a colored rectangle.

GooglePrivacyDlpV2DeidentifyConfig
The configuration that controls how the data will change.

GooglePrivacyDlpV2ListStoredInfoTypesResponse
Response message for ListStoredInfoTypes.

GooglePrivacyDlpV2InfoTypeLimit
Max findings configuration per infoType, per content item or long running DlpJob.

GooglePrivacyDlpV2InspectionRule
A single inspection rule to be applied to infoTypes, specified in InspectionRuleSet.

GooglePrivacyDlpV2CryptoReplaceFfxFpeConfig
Replaces an identifier with a surrogate using Format Preserving Encryption (FPE) with the FFX mode of operation; however when used in the ReidentifyContent API method, it serves the opposite function by reversing the surrogate back into the original identifier. The identifier must be encoded as ASCII. For a given crypto key and context, the same identifier will be replaced with the same surrogate. Identifiers must be at least two characters long. In the case that the identifier is the empty string, it will be skipped. See https://cloud.google.com/dlp/docs/pseudonymization to learn more.

GooglePrivacyDlpV2QuasiId
A column with a semantic tag attached.

GooglePrivacyDlpV2TimespanConfig
Configuration of the timespan of the items to include in scanning. Currently only supported when inspecting Google Cloud Storage and BigQuery.

GooglePrivacyDlpV2UpdateInspectTemplateRequest
Request message for UpdateInspectTemplate.

GooglePrivacyDlpV2KMapEstimationQuasiIdValues
A tuple of values for the quasi-identifier columns.

GooglePrivacyDlpV2ContentItem
Container structure for the content to inspect.

GooglePrivacyDlpV2CryptoDeterministicConfig
Pseudonymization method that generates deterministic encryption for the given input. Outputs a base64 encoded representation of the encrypted output. Uses AES-SIV based on the RFC https://tools.ietf.org/html/rfc5297.

GooglePrivacyDlpV2CryptoKey
This is a data encryption key (DEK) (as opposed to a key encryption key (KEK) stored by KMS). When using KMS to wrap/unwrap DEKs, be sure to set an appropriate IAM policy on the KMS CryptoKey (KEK) to ensure an attacker cannot unwrap the data crypto key.

GooglePrivacyDlpV2ListInspectTemplatesResponse
Response message for ListInspectTemplates.

GooglePrivacyDlpV2RecordKey
Message for a unique key indicating a record that contains a finding.

GooglePrivacyDlpV2AuxiliaryTable
An auxiliary table contains statistical information on the relative frequency of different quasi-identifiers values. It has one or several quasi-identifiers columns, and one column that indicates the relative frequency of each quasi-identifier tuple. If a tuple is present in the data but not in the auxiliary table, the corresponding relative frequency is assumed to be zero (and thus, the tuple is highly reidentifiable).

GooglePrivacyDlpV2ValueFrequency
A value of a field, including its frequency.

GooglePrivacyDlpV2ActivateJobTriggerRequest
Request message for ActivateJobTrigger.

GooglePrivacyDlpV2PublishToPubSub
Publish a message into given Pub/Sub topic when DlpJob has completed. The message contains a single field, DlpJobName, which is equal to the finished job's DlpJob.name. Compatible with: Inspect, Risk

GooglePrivacyDlpV2ListInfoTypesResponse
Response to the ListInfoTypes request.

GooglePrivacyDlpV2Proximity
Message for specifying a window around a finding to apply a detection rule.

GooglePrivacyDlpV2RecordLocation
Location of a finding within a row or record.

GooglePrivacyDlpV2KAnonymityResult
Result of the k-anonymity computation.

GooglePrivacyDlpV2BoundingBox
Bounding box encompassing detected text within an image.

GooglePrivacyDlpV2KMapEstimationResult
Result of the reidentifiability analysis. Note that these results are an estimation, not exact values.

GooglePrivacyDlpV2ImageRedactionConfig
Configuration for determining how redaction of images should occur.

GooglePrivacyDlpV2CategoricalStatsHistogramBucket
There is no detailed description.

GooglePrivacyDlpV2ByteContentItem
Container for bytes to inspect or redact.

GooglePrivacyDlpV2Condition
The field type of value and field do not need to match to be considered equal, but not all comparisons are possible. EQUAL_TO and NOT_EQUAL_TO attempt to compare even with incompatible types, but all other comparisons are invalid with incompatible types. A value of type:

GooglePrivacyDlpV2EntityId
An entity in a dataset is a field or set of fields that correspond to a single person. For example, in medical records the EntityId might be a patient identifier, or for financial records it might be an account identifier. This message is used when generalizations or analysis must take into account that multiple rows correspond to the same entity.

GooglePrivacyDlpV2RedactImageResponse
Results of redacting an image.

GooglePrivacyDlpV2InspectTemplate
The inspectTemplate contains a configuration (set of types of sensitive data to be detected) to be used anywhere you otherwise would normally specify InspectConfig. See https://cloud.google.com/dlp/docs/concepts-templates to learn more.

GooglePrivacyDlpV2QuasiIdField
A quasi-identifier column has a custom_tag, used to know which column in the data corresponds to which column in the statistical model.

GooglePrivacyDlpV2TableLocation
Location of a finding within a table.

GooglePrivacyDlpV2FixedSizeBucketingConfig
Buckets values based on fixed size ranges. The Bucketing transformation can provide all of this functionality, but requires more configuration. This message is provided as a convenience to the user for simple bucketing strategies.

GooglePrivacyDlpV2QuasiIdentifierField
A quasi-identifier column has a custom_tag, used to know which column in the data corresponds to which column in the statistical model.

GooglePrivacyDlpV2UnwrappedCryptoKey
Using raw keys is prone to security risks due to accidentally leaking the key. Choose another type of key if possible.

GooglePrivacyDlpV2ExcludeInfoTypes
List of exclude infoTypes.

GooglePrivacyDlpV2DeidentifyContentRequest
Request to de-identify a list of items.

GooglePrivacyDlpV2Finding
Represents a piece of potentially sensitive content.

GooglePrivacyDlpV2Conditions
A collection of conditions.

GooglePrivacyDlpV2StatisticalTable
An auxiliary table containing statistical information on the relative frequency of different quasi-identifiers values. It has one or several quasi-identifiers columns, and one column that indicates the relative frequency of each quasi-identifier tuple. If a tuple is present in the data but not in the auxiliary table, the corresponding relative frequency is assumed to be zero (and thus, the tuple is highly reidentifiable).

GooglePrivacyDlpV2PrivacyMetric
Privacy metric to compute for reidentification risk analysis.

GooglePrivacyDlpV2BigQueryField
Message defining a field of a BigQuery table.

GooglePrivacyDlpV2InfoType
Type of information detected by the API.

GooglePrivacyDlpV2NumericalStatsResult
Result of the numerical stats computation.

GooglePrivacyDlpV2InfoTypeTransformations
A type of transformation that will scan unstructured text and apply various PrimitiveTransformations to each finding, where the transformation is applied to only values that were identified as a specific info_type.

GooglePrivacyDlpV2CharacterMaskConfig
Partially mask a string by replacing a given number of characters with a fixed character. Masking can start from the beginning or end of the string. This can be used on data of any type (numbers, longs, and so on) and when de-identifying structured data we'll attempt to preserve the original data's type. (This allows you to take a long like 123 and modify it to a string like **3.

GooglePrivacyDlpV2SummaryResult
A collection that informs the user the number of times a particular TransformationResultCode and error details occurred.

GooglePrivacyDlpV2SaveFindings
If set, the detailed findings will be persisted to the specified OutputStorageConfig. Only a single instance of this action can be specified. Compatible with: Inspect, Risk

GooglePrivacyDlpV2DeltaPresenceEstimationQuasiIdValues
A tuple of values for the quasi-identifier columns.

GooglePrivacyDlpV2LDiversityEquivalenceClass
The set of columns' values that share the same ldiversity value.

GooglePrivacyDlpV2ExclusionRule
The rule that specifies conditions when findings of infoTypes specified in InspectionRuleSet are removed from results.

GooglePrivacyDlpV2TransformationSummary
Summary of a single transformation. Only one of 'transformation', 'field_transformation', or 'record_suppress' will be set.

GooglePrivacyDlpV2DateShiftConfig
Shifts dates by random number of days, with option to be consistent for the same context. See https://cloud.google.com/dlp/docs/concepts-date-shifting to learn more.

GooglePrivacyDlpV2RedactConfig
Redact a given value. For example, if used with an InfoTypeTransformation transforming PHONE_NUMBER, and input 'My phone number is 206-555-0123', the output would be 'My phone number is '.

GooglePrivacyDlpV2PublishFindingsToCloudDataCatalog
Publish findings of a DlpJob to Cloud Data Catalog. Labels summarizing the results of the DlpJob will be applied to the entry for the resource scanned in Cloud Data Catalog. Any labels previously written by another DlpJob will be deleted. InfoType naming patterns are strictly enforced when using this feature. Note that the findings will be persisted in Cloud Data Catalog storage and are governed by Data Catalog service-specific policy, see https://cloud.google.com/terms/service-terms Only a single instance of this action can be specified and only allowed if all resources being scanned are BigQuery tables. Compatible with: Inspect

GooglePrivacyDlpV2RequestedOptions
There is no detailed description.

GooglePrivacyDlpV2DatastoreOptions
Options defining a data set within Google Cloud Datastore.

GooglePrivacyDlpV2Range
Generic half-open interval [start, end)

GooglePrivacyDlpV2PrimitiveTransformation
A rule for transforming a value.

GooglePrivacyDlpV2InfoTypeDescription
InfoType description.

GooglePrivacyDlpV2Row
There is no detailed description.

GooglePrivacyDlpV2ListJobTriggersResponse
Response message for ListJobTriggers.

GooglePrivacyDlpV2BigQueryOptions
Options defining BigQuery table and row identifiers.

GooglePrivacyDlpV2UpdateStoredInfoTypeRequest
Request message for UpdateStoredInfoType.

GooglePrivacyDlpV2DetectionRule
Deprecated; use InspectionRuleSet instead. Rule for modifying a CustomInfoType to alter behavior under certain circumstances, depending on the specific details of the rule. Not supported for the surrogate_type custom infoType.

GooglePrivacyDlpV2CreateInspectTemplateRequest
Request message for CreateInspectTemplate.

GooglePrivacyDlpV2StorageConfig
Shared message indicating Cloud storage type.

GooglePrivacyDlpV2ReplaceValueConfig
Replace each input value with a given Value.

GooglePrivacyDlpV2InfoTypeTransformation
A transformation to apply to text that is identified as a specific info_type.

GooglePrivacyDlpV2InspectConfig
Configuration description of the scanning process. When used with redactContent only info_types and min_likelihood are currently used.

GooglePrivacyDlpV2InspectDataSourceDetails
The results of an inspect DataSource job.

GooglePrivacyDlpV2InspectionRuleSet
Rule set for modifying a set of infoTypes to alter behavior under certain circumstances, depending on the specific details of the rules within the set.

GooglePrivacyDlpV2InspectResult
All the findings for a single scanned item.

GooglePrivacyDlpV2DateTime
Message for a date time object. e.g. 2018-01-01, 5th August.

GooglePrivacyDlpV2Color
Represents a color in the RGB color space.

GooglePrivacyDlpV2Action
A task to execute on the completion of a job. See https://cloud.google.com/dlp/docs/concepts-actions to learn more.

GooglePrivacyDlpV2BigQueryKey
Row key for identifying a record in BigQuery table.

GooglePrivacyDlpV2BigQueryTable
Message defining the location of a BigQuery table. A table is uniquely identified by its project_id, dataset_id, and table_name. Within a query a table is often referenced with a string in the format of: <project_id>:<dataset_id>.<table_id> or <project_id>.<dataset_id>.<table_id>.

GooglePrivacyDlpV2UpdateDeidentifyTemplateRequest
Request message for UpdateDeidentifyTemplate.

GooglePrivacyDlpV2BucketingConfig
Generalization function that buckets values based on ranges. The ranges and replacement values are dynamically provided by the user for custom behavior, such as 1-30 -> LOW 31-65 -> MEDIUM 66-100 -> HIGH This can be used on data of type: number, long, string, timestamp. If the bound Value type differs from the type of data being transformed, we will first attempt converting the type of the data to be transformed to match the type of the bound before comparing. See https://cloud.google.com/dlp/docs/concepts-bucketing to learn more.

GooglePrivacyDlpV2CharsToIgnore
Characters to skip when doing deidentification of a value. These will be left alone and skipped.

GooglePrivacyDlpV2PartitionId
Datastore partition ID. A partition ID identifies a grouping of entities. The grouping is always by project and namespace, however the namespace ID may be empty.

GooglePrivacyDlpV2StoredInfoType
StoredInfoType resource message that contains information about the current version and any pending updates.

GooglePrivacyDlpV2ReplaceWithInfoTypeConfig
Replace each matching finding with the name of the info_type.

GooglePrivacyDlpV2AnalyzeDataSourceRiskDetails
Result of a risk analysis operation request.

GooglePrivacyDlpV2CreateJobTriggerRequest
Request message for CreateJobTrigger.

GooglePrivacyDlpV2DocumentLocation
Location of a finding within a document.

GooglePrivacyDlpV2KMapEstimationHistogramBucket
A KMapEstimationHistogramBucket message with the following values: min_anonymity: 3 max_anonymity: 5 frequency: 42 means that there are 42 records whose quasi-identifier values correspond to 3, 4 or 5 people in the overlying population. An important particular case is when min_anonymity = max_anonymity = 1: the frequency field then corresponds to the number of uniquely identifiable records.

GooglePrivacyDlpV2ReidentifyContentResponse
Results of re-identifying a item.

GooglePrivacyDlpV2FieldId
General identifier of a data field in a storage service.

GooglePrivacyDlpV2Bucket
Bucket is represented as a range, along with replacement values.

GooglePrivacyDlpV2KAnonymityHistogramBucket
There is no detailed description.

GooglePrivacyDlpV2LDiversityConfig
l-diversity metric, used for analysis of reidentification risk.

GooglePrivacyDlpV2CloudStoragePath
Message representing a single file or path in Cloud Storage.

GooglePrivacyDlpV2FileSet
Set of files to scan.

GooglePrivacyDlpV2SurrogateType
Message for detecting output from deidentification transformations such as CryptoReplaceFfxFpeConfig. These types of transformations are those that perform pseudonymization, thereby producing a "surrogate" as output. This should be used in conjunction with a field on the transformation such as surrogate_info_type. This CustomInfoType does not support the use of detection_rules.

GooglePrivacyDlpV2LikelihoodAdjustment
Message for specifying an adjustment to the likelihood of a finding as part of a detection rule.

GooglePrivacyDlpV2Error
Details information about an error encountered during job execution or the results of an unsuccessful activation of the JobTrigger. Output only field.

GooglePrivacyDlpV2InspectContentResponse
Results of inspecting an item.

GooglePrivacyDlpV2Table
Structured content to inspect. Up to 50,000 Values per request allowed. See https://cloud.google.com/dlp/docs/inspecting-text#inspecting_a_table to learn more.

GooglePrivacyDlpV2QuoteInfo
Message for infoType-dependent details parsed from quote.

GooglePrivacyDlpV2UpdateJobTriggerRequest
Request message for UpdateJobTrigger.

GooglePrivacyDlpV2OutputStorageConfig
Cloud repository for storing output.

GooglePrivacyDlpV2JobTrigger
Contains a configuration to make dlp api calls on a repeating basis. See https://cloud.google.com/dlp/docs/concepts-job-triggers to learn more.

GooglePrivacyDlpV2StoredType
A reference to a StoredInfoType to use with scanning.

GooglePrivacyDlpV2Schedule
Schedule for triggeredJobs.

GooglePrivacyDlpV2FindingLimits
There is no detailed description.

GooglePrivacyDlpV2KmsWrappedCryptoKey
Include to use an existing data crypto key wrapped by KMS. The wrapped key must be a 128/192/256 bit key. Authorization requires the following IAM permissions when sending a request to perform a crypto transformation using a kms-wrapped crypto key: dlp.kms.encrypt

GooglePrivacyDlpV2LargeCustomDictionaryStats
Summary statistics of a custom dictionary.

GooglePrivacyDlpV2ImageLocation
Location of the finding within an image.

GooglePrivacyDlpV2Trigger
What event needs to occur for a new job to be started.

GooglePrivacyDlpV2WordList
Message defining a list of words or phrases to search for in the data.

GooglePrivacyDlpV2StoredInfoTypeStats
Statistics for a StoredInfoType.

GooglePrivacyDlpV2TimeZone
There is no detailed description.

GooglePrivacyDlpV2CryptoHashConfig
Pseudonymization method that generates surrogates via cryptographic hashing. Uses SHA-256. The key size must be either 32 or 64 bytes. Outputs a base64 encoded representation of the hashed output (for example, L7k0BHmF1ha5U3NfGykjro4xWi1MPVQPjhMAZbSV9mM=). Currently, only string and integer values can be hashed. See https://cloud.google.com/dlp/docs/pseudonymization to learn more.

GooglePrivacyDlpV2ListDeidentifyTemplatesResponse
Response message for ListDeidentifyTemplates.

GooglePrivacyDlpV2Regex
Message defining a custom regular expression.

GooglePrivacyDlpV2FieldTransformation
The transformation to apply to the field.

GooglePrivacyDlpV2RecordCondition
A condition for determining whether a transformation should be applied to a field.

GooglePrivacyDlpV2DeltaPresenceEstimationHistogramBucket
A DeltaPresenceEstimationHistogramBucket message with the following values: min_probability: 0.1 max_probability: 0.2 frequency: 42 means that there are 42 records for which δ is in [0.1, 0.2). An important particular case is when min_probability = max_probability = 1: then, every individual who shares this quasi-identifier combination is in the dataset.

GooglePrivacyDlpV2Location
Specifies the location of the finding.

GooglePrivacyDlpV2ContentLocation
Findings container location data.

GooglePrivacyDlpV2ReidentifyContentRequest
Request to re-identify an item.

GooglePrivacyDlpV2KindExpression
A representation of a Datastore kind.

GooglePrivacyDlpV2KAnonymityConfig
k-anonymity metric, used for analysis of reidentification risk.

GooglePrivacyDlpV2CloudStorageFileSet
Message representing a set of files in Cloud Storage.

GooglePrivacyDlpV2DeltaPresenceEstimationConfig
δ-presence metric, used to estimate how likely it is for an attacker to figure out that one given individual appears in a de-identified dataset. Similarly to the k-map metric, we cannot compute δ-presence exactly without knowing the attack dataset, so we use a statistical model instead.

GooglePrivacyDlpV2StoredInfoTypeVersion
Version of a StoredInfoType, including the configuration used to build it, create timestamp, and current state.

GooglePrivacyDlpV2KMapEstimationConfig
Reidentifiability metric. This corresponds to a risk model similar to what is called "journalist risk" in the literature, except the attack dataset is statistically modeled instead of being perfectly known. This can be done using publicly available data (like the US Census), or using a custom statistical model (indicated as one or several BigQuery tables), or by extrapolating from the distribution of values in the input dataset. A column with a semantic tag attached.

GooglePrivacyDlpV2DeidentifyContentResponse
Results of de-identifying a ContentItem.

GooglePrivacyDlpV2Value
Set of primitive values supported by the system. Note that for the purposes of inspection or transformation, the number of bytes considered to comprise a 'Value' is based on its representation as a UTF-8 encoded string. For example, if 'integer_value' is set to 123456789, the number of bytes would be counted as 9, even though an int64 only holds up to 8 bytes of data.

GooglePrivacyDlpV2HotwordRule
The rule that adjusts the likelihood of findings within a certain proximity of hotwords.

GooglePrivacyDlpV2KAnonymityEquivalenceClass
The set of columns' values that share the same ldiversity value

GooglePrivacyDlpV2DatastoreKey
Record key for a finding in Cloud Datastore.

GooglePrivacyDlpV2CreateDeidentifyTemplateRequest
Request message for CreateDeidentifyTemplate.

GooglePrivacyDlpV2CreateStoredInfoTypeRequest
Request message for CreateStoredInfoType.

GooglePrivacyDlpV2Expressions
An expression, consisting or an operator and conditions.

GooglePrivacyDlpV2CloudStorageOptions
Options defining a file or a set of files within a Google Cloud Storage bucket.

GooglePrivacyDlpV2CloudStorageRegexFileSet
Message representing a set of files in a Cloud Storage bucket. Regular expressions are used to allow fine-grained control over which files in the bucket to include.

GooglePrivacyDlpV2RecordSuppression
Configuration to suppress records whose suppression conditions evaluate to true.

GooglePrivacyDlpV2PublishSummaryToCscc
Publish the result summary of a DlpJob to the Cloud Security Command Center (CSCC Alpha). This action is only available for projects which are parts of an organization and whitelisted for the alpha Cloud Security Command Center. The action will publish count of finding instances and their info types. The summary of findings will be persisted in CSCC and are governed by CSCC service-specific policy, see https://cloud.google.com/terms/service-terms Only a single instance of this action can be specified. Compatible with: Inspect

GooglePrivacyDlpV2TransformationOverview
Overview of the modifications that occurred.

GooglePrivacyDlpV2CategoricalStatsResult
Result of the categorical stats computation.

GooglePrivacyDlpV2LargeCustomDictionaryConfig
Configuration for a custom dictionary created from a data source of any size up to the maximum size defined in the limits page. The artifacts of dictionary creation are stored in the specified Google Cloud Storage location. Consider using CustomInfoType.Dictionary for smaller dictionaries that satisfy the size requirements.

GooglePrivacyDlpV2RiskAnalysisJobConfig
Configuration for a risk analysis job. See https://cloud.google.com/dlp/docs/concepts-risk-analysis to learn more.

GooglePrivacyDlpV2DlpJob
Combines all of the information about a DLP job.

GooglePrivacyDlpV2ListDlpJobsResponse
The response message for listing DLP jobs.

GooglePrivacyDlpV2Dictionary
Custom information type based on a dictionary of words or phrases. This can be used to match sensitive information specific to the data, such as a list of employee IDs or job titles.

GooglePrivacyDlpV2InfoTypeStats
Statistics regarding a specific InfoType.

GooglePrivacyDlpV2TimePartConfig
For use with Date, Timestamp, and TimeOfDay, extract or preserve a portion of the value.

GooglePrivacyDlpV2CancelDlpJobRequest
The request message for canceling a DLP job.

GooglePrivacyDlpV2Result
All result fields mentioned below are updated while the job is processing.

GooglePrivacyDlpV2RecordTransformations
A type of transformation that is applied over structured data such as a table.

GooglePrivacyDlpV2TaggedField
There is no detailed description.

GooglePrivacyDlpV2TransientCryptoKey
Use this to have a random data crypto key generated. It will be discarded after the request finishes.

GooglePrivacyDlpV2LDiversityResult
Result of the l-diversity computation.

GooglePrivacyDlpV2LDiversityHistogramBucket
There is no detailed description.

GooglePrivacyDlpV2StoredInfoTypeConfig
Configuration for a StoredInfoType.

GooglePrivacyDlpV2CategoricalStatsConfig
Compute numerical stats over an individual column, including number of distinct values and value count distribution.

GooglePrivacyDlpV2DeidentifyTemplate
The DeidentifyTemplates contains instructions on how to deidentify content. See https://cloud.google.com/dlp/docs/concepts-templates to learn more.

GooglePrivacyDlpV2DeltaPresenceEstimationResult
Result of the δ-presence computation. Note that these results are an estimation, not exact values.

GooglePrivacyDlpV2Key
A unique identifier for a Datastore entity. If a key's partition ID or any of its path kinds or names are reserved/read-only, the key is reserved/read-only. A reserved/read-only key is forbidden in certain documented contexts.

GooglePrivacyDlpV2JobNotificationEmails
Enable email notification to project owners and editors on jobs's completion/failure.

GooglePrivacyDlpV2PathElement
A (kind, ID/name) pair used to construct a key path.

GooglePrivacyDlpV2CustomInfoType
Custom information type provided by the user. Used to find domain-specific sensitive information configurable to the data in question.

GooglePrivacyDlpV2NumericalStatsConfig
Compute numerical stats over an individual column, including min, max, and quantiles.

GooglePrivacyDlpV2InspectContentRequest
Request to search for potentially sensitive info in a ContentItem.

GoogleProtobufEmpty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance:

GoogleRpcStatus
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details.

GoogleTypeDate
Represents a whole or partial calendar date, e.g. a birthday. The time of day and time zone are either specified elsewhere or are not significant. The date is relative to the Proleptic Gregorian Calendar. This can represent:

GoogleTypeTimeOfDay
Represents a time of day. The date and time zone are either not significant or are specified elsewhere. An API may choose to allow leap seconds. Related types are google.type.Date and google.protobuf.Timestamp.

InfoTypeListCall
Returns a list of the sensitive information types that the DLP API supports. See https://cloud.google.com/dlp/docs/infotypes-reference to learn more.

InfoTypeMethods
A builder providing access to all methods supported on infoType resources. It is not used directly, but through the DLP hub.

JsonServerError
A utility type which can decode a server response that indicates error

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

OrganizationDeidentifyTemplateCreateCall
Creates a DeidentifyTemplate for re-using frequently used configuration for de-identifying content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

OrganizationDeidentifyTemplateDeleteCall
Deletes a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

OrganizationDeidentifyTemplateGetCall
Gets a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

OrganizationDeidentifyTemplateListCall
Lists DeidentifyTemplates. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

OrganizationDeidentifyTemplatePatchCall
Updates the DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

OrganizationInspectTemplateCreateCall
Creates an InspectTemplate for re-using frequently used configuration for inspecting content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

OrganizationInspectTemplateDeleteCall
Deletes an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

OrganizationInspectTemplateGetCall
Gets an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

OrganizationInspectTemplateListCall
Lists InspectTemplates. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

OrganizationInspectTemplatePatchCall
Updates the InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

OrganizationMethods
A builder providing access to all methods supported on organization resources. It is not used directly, but through the DLP hub.

OrganizationStoredInfoTypeCreateCall
Creates a pre-built stored infoType to be used for inspection. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

OrganizationStoredInfoTypeDeleteCall
Deletes a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

OrganizationStoredInfoTypeGetCall
Gets a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

OrganizationStoredInfoTypeListCall
Lists stored infoTypes. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

OrganizationStoredInfoTypePatchCall
Updates the stored infoType by creating a new version. The existing version will continue to be used until the new version is ready. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

ProjectContentDeidentifyCall
De-identifies potentially sensitive info from a ContentItem. This method has limits on input size and output size. See https://cloud.google.com/dlp/docs/deidentify-sensitive-data to learn more.

ProjectContentInspectCall
Finds potentially sensitive info in content. This method has limits on input size, processing time, and output size.

ProjectContentReidentifyCall
Re-identifies content that has been de-identified. See https://cloud.google.com/dlp/docs/pseudonymization#re-identification_in_free_text_code_example to learn more.

ProjectDeidentifyTemplateCreateCall
Creates a DeidentifyTemplate for re-using frequently used configuration for de-identifying content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

ProjectDeidentifyTemplateDeleteCall
Deletes a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

ProjectDeidentifyTemplateGetCall
Gets a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

ProjectDeidentifyTemplateListCall
Lists DeidentifyTemplates. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

ProjectDeidentifyTemplatePatchCall
Updates the DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

ProjectDlpJobCancelCall
Starts asynchronous cancellation on a long-running DlpJob. The server makes a best effort to cancel the DlpJob, but success is not guaranteed. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

ProjectDlpJobCreateCall
Creates a new job to inspect storage or calculate risk metrics. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

ProjectDlpJobDeleteCall
Deletes a long-running DlpJob. This method indicates that the client is no longer interested in the DlpJob result. The job will be cancelled if possible. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

ProjectDlpJobGetCall
Gets the latest state of a long-running DlpJob. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

ProjectDlpJobListCall
Lists DlpJobs that match the specified filter in the request. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

ProjectImageRedactCall
Redacts potentially sensitive info from an image. This method has limits on input size, processing time, and output size. See https://cloud.google.com/dlp/docs/redacting-sensitive-data-images to learn more.

ProjectInspectTemplateCreateCall
Creates an InspectTemplate for re-using frequently used configuration for inspecting content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

ProjectInspectTemplateDeleteCall
Deletes an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

ProjectInspectTemplateGetCall
Gets an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

ProjectInspectTemplateListCall
Lists InspectTemplates. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

ProjectInspectTemplatePatchCall
Updates the InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

ProjectJobTriggerActivateCall
Activate a job trigger. Causes the immediate execute of a trigger instead of waiting on the trigger event to occur.

ProjectJobTriggerCreateCall
Creates a job trigger to run DLP actions such as scanning storage for sensitive information on a set schedule. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

ProjectJobTriggerDeleteCall
Deletes a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

ProjectJobTriggerGetCall
Gets a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

ProjectJobTriggerListCall
Lists job triggers. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

ProjectJobTriggerPatchCall
Updates a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

ProjectLocationContentDeidentifyCall
De-identifies potentially sensitive info from a ContentItem. This method has limits on input size and output size. See https://cloud.google.com/dlp/docs/deidentify-sensitive-data to learn more.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the DLP hub.

ProjectStoredInfoTypeCreateCall
Creates a pre-built stored infoType to be used for inspection. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

ProjectStoredInfoTypeDeleteCall
Deletes a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

ProjectStoredInfoTypeGetCall
Gets a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

ProjectStoredInfoTypeListCall
Lists stored infoTypes. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

ProjectStoredInfoTypePatchCall
Updates the stored infoType by creating a new version. The existing version will continue to be used until the new version is ready. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

RangeResponseHeader

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

ServerError

ServerMessage

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.