Module google_dlp2::api

Structs

DLP
Central instance to access all DLP related resource activities

GooglePrivacyDlpV2Action
A task to execute on the completion of a job. See https://cloud.google.com/dlp/docs/concepts-actions to learn more.

GooglePrivacyDlpV2ActivateJobTriggerRequest
Request message for ActivateJobTrigger.

GooglePrivacyDlpV2AnalyzeDataSourceRiskDetails
Result of a risk analysis operation request.

GooglePrivacyDlpV2AuxiliaryTable
An auxiliary table contains statistical information on the relative frequency of different quasi-identifiers values. It has one or several quasi-identifiers columns, and one column that indicates the relative frequency of each quasi-identifier tuple. If a tuple is present in the data but not in the auxiliary table, the corresponding relative frequency is assumed to be zero (and thus, the tuple is highly reidentifiable).

GooglePrivacyDlpV2BigQueryField
Message defining a field of a BigQuery table.

GooglePrivacyDlpV2BigQueryKey
Row key for identifying a record in BigQuery table.

GooglePrivacyDlpV2BigQueryOptions
Options defining BigQuery table and row identifiers.

GooglePrivacyDlpV2BigQueryTable
Message defining the location of a BigQuery table. A table is uniquely identified by its project_id, dataset_id, and table_name. Within a query a table is often referenced with a string in the format of: :. or ...

GooglePrivacyDlpV2BoundingBox
Bounding box encompassing detected text within an image.

GooglePrivacyDlpV2Bucket
Bucket is represented as a range, along with replacement values.

GooglePrivacyDlpV2BucketingConfig
Generalization function that buckets values based on ranges. The ranges and replacement values are dynamically provided by the user for custom behavior, such as 1-30 -> LOW 31-65 -> MEDIUM 66-100 -> HIGH This can be used on data of type: number, long, string, timestamp. If the bound Value type differs from the type of data being transformed, we will first attempt converting the type of the data to be transformed to match the type of the bound before comparing. See https://cloud.google.com/dlp/docs/concepts-bucketing to learn more.

GooglePrivacyDlpV2ByteContentItem
Container for bytes to inspect or redact.

GooglePrivacyDlpV2CancelDlpJobRequest
The request message for canceling a DLP job.

GooglePrivacyDlpV2CategoricalStatsConfig
Compute numerical stats over an individual column, including number of distinct values and value count distribution.

GooglePrivacyDlpV2CategoricalStatsHistogramBucket
Histogram of value frequencies in the column.

GooglePrivacyDlpV2CategoricalStatsResult
Result of the categorical stats computation.

GooglePrivacyDlpV2CharacterMaskConfig
Partially mask a string by replacing a given number of characters with a fixed character. Masking can start from the beginning or end of the string. This can be used on data of any type (numbers, longs, and so on) and when de-identifying structured data we'll attempt to preserve the original data's type. (This allows you to take a long like 123 and modify it to a string like **3.

GooglePrivacyDlpV2CharsToIgnore
Characters to skip when doing deidentification of a value. These will be left alone and skipped.

GooglePrivacyDlpV2CloudStorageFileSet
Message representing a set of files in Cloud Storage.

GooglePrivacyDlpV2CloudStorageOptions
Options defining a file or a set of files within a Google Cloud Storage bucket.

GooglePrivacyDlpV2CloudStoragePath
Message representing a single file or path in Cloud Storage.

GooglePrivacyDlpV2CloudStorageRegexFileSet
Message representing a set of files in a Cloud Storage bucket. Regular expressions are used to allow fine-grained control over which files in the bucket to include. Included files are those that match at least one item in include_regex and do not match any items in exclude_regex. Note that a file that matches items from both lists will not be included. For a match to occur, the entire file path (i.e., everything in the url after the bucket name) must match the regular expression. For example, given the input {bucket_name: "mybucket", include_regex: ["directory1/.*"], exclude_regex: ["directory1/excluded.*"]}: * gs://mybucket/directory1/myfile will be included * gs://mybucket/directory1/directory2/myfile will be included (.* matches across /) * gs://mybucket/directory0/directory1/myfile will not be included (the full path doesn't match any items in include_regex) * gs://mybucket/directory1/excludedfile will not be included (the path matches an item in exclude_regex) If include_regex is left empty, it will match all files by default (this is equivalent to setting include_regex: [".*"]). Some other common use cases: * {bucket_name: "mybucket", exclude_regex: [".*\.pdf"]} will include all files in mybucket except for .pdf files * {bucket_name: "mybucket", include_regex: ["directory/[^/]+"]} will include all files directly under gs://mybucket/directory/, without matching across /

GooglePrivacyDlpV2Color
Represents a color in the RGB color space.

GooglePrivacyDlpV2Condition
The field type of value and field do not need to match to be considered equal, but not all comparisons are possible. EQUAL_TO and NOT_EQUAL_TO attempt to compare even with incompatible types, but all other comparisons are invalid with incompatible types. A value of type: - string can be compared against all other types - boolean can only be compared against other booleans - integer can be compared against doubles or a string if the string value can be parsed as an integer. - double can be compared against integers or a string if the string can be parsed as a double. - Timestamp can be compared against strings in RFC 3339 date string format. - TimeOfDay can be compared against timestamps and strings in the format of 'HH:mm:ss'. If we fail to compare do to type mismatch, a warning will be given and the condition will evaluate to false.

GooglePrivacyDlpV2Conditions
A collection of conditions.

GooglePrivacyDlpV2Container
Represents a container that may contain DLP findings. Examples of a container include a file, table, or database record.

GooglePrivacyDlpV2ContentItem
Container structure for the content to inspect.

GooglePrivacyDlpV2ContentLocation
Precise location of the finding within a document, record, image, or metadata container.

GooglePrivacyDlpV2CreateDeidentifyTemplateRequest
Request message for CreateDeidentifyTemplate.

GooglePrivacyDlpV2CreateDlpJobRequest
Request message for CreateDlpJobRequest. Used to initiate long running jobs such as calculating risk metrics or inspecting Google Cloud Storage.

GooglePrivacyDlpV2CreateInspectTemplateRequest
Request message for CreateInspectTemplate.

GooglePrivacyDlpV2CreateJobTriggerRequest
Request message for CreateJobTrigger.

GooglePrivacyDlpV2CreateStoredInfoTypeRequest
Request message for CreateStoredInfoType.

GooglePrivacyDlpV2CryptoDeterministicConfig
Pseudonymization method that generates deterministic encryption for the given input. Outputs a base64 encoded representation of the encrypted output. Uses AES-SIV based on the RFC https://tools.ietf.org/html/rfc5297.

GooglePrivacyDlpV2CryptoHashConfig
Pseudonymization method that generates surrogates via cryptographic hashing. Uses SHA-256. The key size must be either 32 or 64 bytes. Outputs a base64 encoded representation of the hashed output (for example, L7k0BHmF1ha5U3NfGykjro4xWi1MPVQPjhMAZbSV9mM=). Currently, only string and integer values can be hashed. See https://cloud.google.com/dlp/docs/pseudonymization to learn more.

GooglePrivacyDlpV2CryptoKey
This is a data encryption key (DEK) (as opposed to a key encryption key (KEK) stored by KMS). When using KMS to wrap/unwrap DEKs, be sure to set an appropriate IAM policy on the KMS CryptoKey (KEK) to ensure an attacker cannot unwrap the data crypto key.

GooglePrivacyDlpV2CryptoReplaceFfxFpeConfig
Replaces an identifier with a surrogate using Format Preserving Encryption (FPE) with the FFX mode of operation; however when used in the ReidentifyContent API method, it serves the opposite function by reversing the surrogate back into the original identifier. The identifier must be encoded as ASCII. For a given crypto key and context, the same identifier will be replaced with the same surrogate. Identifiers must be at least two characters long. In the case that the identifier is the empty string, it will be skipped. See https://cloud.google.com/dlp/docs/pseudonymization to learn more. Note: We recommend using CryptoDeterministicConfig for all use cases which do not require preserving the input alphabet space and size, plus warrant referential integrity.

GooglePrivacyDlpV2CustomInfoType
Custom information type provided by the user. Used to find domain-specific sensitive information configurable to the data in question.

GooglePrivacyDlpV2DatastoreKey
Record key for a finding in Cloud Datastore.

GooglePrivacyDlpV2DatastoreOptions
Options defining a data set within Google Cloud Datastore.

GooglePrivacyDlpV2DateShiftConfig
Shifts dates by random number of days, with option to be consistent for the same context. See https://cloud.google.com/dlp/docs/concepts-date-shifting to learn more.

GooglePrivacyDlpV2DateTime
Message for a date time object. e.g. 2018-01-01, 5th August.

GooglePrivacyDlpV2DeidentifyConfig
The configuration that controls how the data will change.

GooglePrivacyDlpV2DeidentifyContentRequest
Request to de-identify a list of items.

GooglePrivacyDlpV2DeidentifyContentResponse
Results of de-identifying a ContentItem.

GooglePrivacyDlpV2DeidentifyTemplate
DeidentifyTemplates contains instructions on how to de-identify content. See https://cloud.google.com/dlp/docs/concepts-templates to learn more.

GooglePrivacyDlpV2DeltaPresenceEstimationConfig
δ-presence metric, used to estimate how likely it is for an attacker to figure out that one given individual appears in a de-identified dataset. Similarly to the k-map metric, we cannot compute δ-presence exactly without knowing the attack dataset, so we use a statistical model instead.

GooglePrivacyDlpV2DeltaPresenceEstimationHistogramBucket
A DeltaPresenceEstimationHistogramBucket message with the following values: min_probability: 0.1 max_probability: 0.2 frequency: 42 means that there are 42 records for which δ is in [0.1, 0.2). An important particular case is when min_probability = max_probability = 1: then, every individual who shares this quasi-identifier combination is in the dataset.

GooglePrivacyDlpV2DeltaPresenceEstimationQuasiIdValues
A tuple of values for the quasi-identifier columns.

GooglePrivacyDlpV2DeltaPresenceEstimationResult
Result of the δ-presence computation. Note that these results are an estimation, not exact values.

GooglePrivacyDlpV2DetectionRule
Deprecated; use InspectionRuleSet instead. Rule for modifying a CustomInfoType to alter behavior under certain circumstances, depending on the specific details of the rule. Not supported for the surrogate_type custom infoType.

GooglePrivacyDlpV2Dictionary
Custom information type based on a dictionary of words or phrases. This can be used to match sensitive information specific to the data, such as a list of employee IDs or job titles. Dictionary words are case-insensitive and all characters other than letters and digits in the unicode Basic Multilingual Plane will be replaced with whitespace when scanning for matches, so the dictionary phrase "Sam Johnson" will match all three phrases "sam johnson", "Sam, Johnson", and "Sam (Johnson)". Additionally, the characters surrounding any match must be of a different type than the adjacent characters within the word, so letters must be next to non-letters and digits next to non-digits. For example, the dictionary word "jen" will match the first three letters of the text "jen123" but will return no matches for "jennifer". Dictionary words containing a large number of characters that are not letters or digits may result in unexpected findings because such characters are treated as whitespace. The limits page contains details about the size limits of dictionaries. For dictionaries that do not fit within these constraints, consider using LargeCustomDictionaryConfig in the StoredInfoType API.

GooglePrivacyDlpV2DlpJob
Combines all of the information about a DLP job.

GooglePrivacyDlpV2DocumentLocation
Location of a finding within a document.

GooglePrivacyDlpV2EntityId
An entity in a dataset is a field or set of fields that correspond to a single person. For example, in medical records the EntityId might be a patient identifier, or for financial records it might be an account identifier. This message is used when generalizations or analysis must take into account that multiple rows correspond to the same entity.

GooglePrivacyDlpV2Error
Details information about an error encountered during job execution or the results of an unsuccessful activation of the JobTrigger.

GooglePrivacyDlpV2ExcludeInfoTypes
List of exclude infoTypes.

GooglePrivacyDlpV2ExclusionRule
The rule that specifies conditions when findings of infoTypes specified in InspectionRuleSet are removed from results.

GooglePrivacyDlpV2Expressions
An expression, consisting or an operator and conditions.

GooglePrivacyDlpV2FieldId
General identifier of a data field in a storage service.

GooglePrivacyDlpV2FieldTransformation
The transformation to apply to the field.

GooglePrivacyDlpV2FileSet
Set of files to scan.

GooglePrivacyDlpV2Finding
Represents a piece of potentially sensitive content.

GooglePrivacyDlpV2FindingLimits
Configuration to control the number of findings returned. Cannot be set if de-identification is requested.

GooglePrivacyDlpV2FinishDlpJobRequest
The request message for finishing a DLP hybrid job.

GooglePrivacyDlpV2FixedSizeBucketingConfig
Buckets values based on fixed size ranges. The Bucketing transformation can provide all of this functionality, but requires more configuration. This message is provided as a convenience to the user for simple bucketing strategies. The transformed value will be a hyphenated string of {lower_bound}-{upper_bound}, i.e if lower_bound = 10 and upper_bound = 20 all values that are within this bucket will be replaced with "10-20". This can be used on data of type: double, long. If the bound Value type differs from the type of data being transformed, we will first attempt converting the type of the data to be transformed to match the type of the bound before comparing. See https://cloud.google.com/dlp/docs/concepts-bucketing to learn more.

GooglePrivacyDlpV2HotwordRule
The rule that adjusts the likelihood of findings within a certain proximity of hotwords.

GooglePrivacyDlpV2HybridContentItem
An individual hybrid item to inspect. Will be stored temporarily during processing.

GooglePrivacyDlpV2HybridFindingDetails
Populate to associate additional data with each finding.

GooglePrivacyDlpV2HybridInspectDlpJobRequest
Request to search for potentially sensitive info in a custom location.

GooglePrivacyDlpV2HybridInspectJobTriggerRequest
Request to search for potentially sensitive info in a custom location.

GooglePrivacyDlpV2HybridInspectResponse
Quota exceeded errors will be thrown once quota has been met.

GooglePrivacyDlpV2HybridInspectStatistics
Statistics related to processing hybrid inspect requests.

GooglePrivacyDlpV2HybridOptions
Configuration to control jobs where the content being inspected is outside of Google Cloud Platform.

GooglePrivacyDlpV2ImageLocation
Location of the finding within an image.

GooglePrivacyDlpV2ImageRedactionConfig
Configuration for determining how redaction of images should occur.

GooglePrivacyDlpV2InfoType
Type of information detected by the API.

GooglePrivacyDlpV2InfoTypeDescription
InfoType description.

GooglePrivacyDlpV2InfoTypeLimit
Max findings configuration per infoType, per content item or long running DlpJob.

GooglePrivacyDlpV2InfoTypeStats
Statistics regarding a specific InfoType.

GooglePrivacyDlpV2InfoTypeTransformation
A transformation to apply to text that is identified as a specific info_type.

GooglePrivacyDlpV2InfoTypeTransformations
A type of transformation that will scan unstructured text and apply various PrimitiveTransformations to each finding, where the transformation is applied to only values that were identified as a specific info_type.

GooglePrivacyDlpV2InspectConfig
Configuration description of the scanning process. When used with redactContent only info_types and min_likelihood are currently used.

GooglePrivacyDlpV2InspectContentRequest
Request to search for potentially sensitive info in a ContentItem.

GooglePrivacyDlpV2InspectContentResponse
Results of inspecting an item.

GooglePrivacyDlpV2InspectDataSourceDetails
The results of an inspect DataSource job.

GooglePrivacyDlpV2InspectJobConfig
Controls what and how to inspect for findings.

GooglePrivacyDlpV2InspectResult
All the findings for a single scanned item.

GooglePrivacyDlpV2InspectTemplate
The inspectTemplate contains a configuration (set of types of sensitive data to be detected) to be used anywhere you otherwise would normally specify InspectConfig. See https://cloud.google.com/dlp/docs/concepts-templates to learn more.

GooglePrivacyDlpV2InspectionRule
A single inspection rule to be applied to infoTypes, specified in InspectionRuleSet.

GooglePrivacyDlpV2InspectionRuleSet
Rule set for modifying a set of infoTypes to alter behavior under certain circumstances, depending on the specific details of the rules within the set.

GooglePrivacyDlpV2JobNotificationEmails
Enable email notification to project owners and editors on jobs's completion/failure.

GooglePrivacyDlpV2JobTrigger
Contains a configuration to make dlp api calls on a repeating basis. See https://cloud.google.com/dlp/docs/concepts-job-triggers to learn more.

GooglePrivacyDlpV2KAnonymityConfig
k-anonymity metric, used for analysis of reidentification risk.

GooglePrivacyDlpV2KAnonymityEquivalenceClass
The set of columns' values that share the same ldiversity value

GooglePrivacyDlpV2KAnonymityHistogramBucket
Histogram of k-anonymity equivalence classes.

GooglePrivacyDlpV2KAnonymityResult
Result of the k-anonymity computation.

GooglePrivacyDlpV2KMapEstimationConfig
Reidentifiability metric. This corresponds to a risk model similar to what is called "journalist risk" in the literature, except the attack dataset is statistically modeled instead of being perfectly known. This can be done using publicly available data (like the US Census), or using a custom statistical model (indicated as one or several BigQuery tables), or by extrapolating from the distribution of values in the input dataset.

GooglePrivacyDlpV2KMapEstimationHistogramBucket
A KMapEstimationHistogramBucket message with the following values: min_anonymity: 3 max_anonymity: 5 frequency: 42 means that there are 42 records whose quasi-identifier values correspond to 3, 4 or 5 people in the overlying population. An important particular case is when min_anonymity = max_anonymity = 1: the frequency field then corresponds to the number of uniquely identifiable records.

GooglePrivacyDlpV2KMapEstimationQuasiIdValues
A tuple of values for the quasi-identifier columns.

GooglePrivacyDlpV2KMapEstimationResult
Result of the reidentifiability analysis. Note that these results are an estimation, not exact values.

GooglePrivacyDlpV2Key
A unique identifier for a Datastore entity. If a key's partition ID or any of its path kinds or names are reserved/read-only, the key is reserved/read-only. A reserved/read-only key is forbidden in certain documented contexts.

GooglePrivacyDlpV2KindExpression
A representation of a Datastore kind.

GooglePrivacyDlpV2KmsWrappedCryptoKey
Include to use an existing data crypto key wrapped by KMS. The wrapped key must be a 128/192/256 bit key. Authorization requires the following IAM permissions when sending a request to perform a crypto transformation using a kms-wrapped crypto key: dlp.kms.encrypt

GooglePrivacyDlpV2LDiversityConfig
l-diversity metric, used for analysis of reidentification risk.

GooglePrivacyDlpV2LDiversityEquivalenceClass
The set of columns' values that share the same ldiversity value.

GooglePrivacyDlpV2LDiversityHistogramBucket
Histogram of l-diversity equivalence class sensitive value frequencies.

GooglePrivacyDlpV2LDiversityResult
Result of the l-diversity computation.

GooglePrivacyDlpV2LargeCustomDictionaryConfig
Configuration for a custom dictionary created from a data source of any size up to the maximum size defined in the limits page. The artifacts of dictionary creation are stored in the specified Google Cloud Storage location. Consider using CustomInfoType.Dictionary for smaller dictionaries that satisfy the size requirements.

GooglePrivacyDlpV2LargeCustomDictionaryStats
Summary statistics of a custom dictionary.

GooglePrivacyDlpV2LeaveUntransformed
Skips the data without modifying it if the requested transformation would cause an error. For example, if a DateShift transformation were applied an an IP address, this mode would leave the IP address unchanged in the response.

GooglePrivacyDlpV2LikelihoodAdjustment
Message for specifying an adjustment to the likelihood of a finding as part of a detection rule.

GooglePrivacyDlpV2ListDeidentifyTemplatesResponse
Response message for ListDeidentifyTemplates.

GooglePrivacyDlpV2ListDlpJobsResponse
The response message for listing DLP jobs.

GooglePrivacyDlpV2ListInfoTypesResponse
Response to the ListInfoTypes request.

GooglePrivacyDlpV2ListInspectTemplatesResponse
Response message for ListInspectTemplates.

GooglePrivacyDlpV2ListJobTriggersResponse
Response message for ListJobTriggers.

GooglePrivacyDlpV2ListStoredInfoTypesResponse
Response message for ListStoredInfoTypes.

GooglePrivacyDlpV2Location
Specifies the location of the finding.

GooglePrivacyDlpV2Manual
Job trigger option for hybrid jobs. Jobs must be manually created and finished.

GooglePrivacyDlpV2MetadataLocation
Metadata Location

GooglePrivacyDlpV2NumericalStatsConfig
Compute numerical stats over an individual column, including min, max, and quantiles.

GooglePrivacyDlpV2NumericalStatsResult
Result of the numerical stats computation.

GooglePrivacyDlpV2OutputStorageConfig
Cloud repository for storing output.

GooglePrivacyDlpV2PartitionId
Datastore partition ID. A partition ID identifies a grouping of entities. The grouping is always by project and namespace, however the namespace ID may be empty. A partition ID contains several dimensions: project ID and namespace ID.

GooglePrivacyDlpV2PathElement
A (kind, ID/name) pair used to construct a key path. If either name or ID is set, the element is complete. If neither is set, the element is incomplete.

GooglePrivacyDlpV2PrimitiveTransformation
A rule for transforming a value.

GooglePrivacyDlpV2PrivacyMetric
Privacy metric to compute for reidentification risk analysis.

GooglePrivacyDlpV2Proximity
Message for specifying a window around a finding to apply a detection rule.

GooglePrivacyDlpV2PublishFindingsToCloudDataCatalog
Publish findings of a DlpJob to Cloud Data Catalog. Labels summarizing the results of the DlpJob will be applied to the entry for the resource scanned in Cloud Data Catalog. Any labels previously written by another DlpJob will be deleted. InfoType naming patterns are strictly enforced when using this feature. Note that the findings will be persisted in Cloud Data Catalog storage and are governed by Data Catalog service-specific policy, see https://cloud.google.com/terms/service-terms Only a single instance of this action can be specified and only allowed if all resources being scanned are BigQuery tables. Compatible with: Inspect

GooglePrivacyDlpV2PublishSummaryToCscc
Publish the result summary of a DlpJob to the Cloud Security Command Center (CSCC Alpha). This action is only available for projects which are parts of an organization and whitelisted for the alpha Cloud Security Command Center. The action will publish count of finding instances and their info types. The summary of findings will be persisted in CSCC and are governed by CSCC service-specific policy, see https://cloud.google.com/terms/service-terms Only a single instance of this action can be specified. Compatible with: Inspect

GooglePrivacyDlpV2PublishToPubSub
Publish a message into given Pub/Sub topic when DlpJob has completed. The message contains a single field, DlpJobName, which is equal to the finished job's DlpJob.name. Compatible with: Inspect, Risk

GooglePrivacyDlpV2PublishToStackdriver
Enable Stackdriver metric dlp.googleapis.com/finding_count. This will publish a metric to stack driver on each infotype requested and how many findings were found for it. CustomDetectors will be bucketed as 'Custom' under the Stackdriver label 'info_type'.

GooglePrivacyDlpV2QuasiId
A column with a semantic tag attached.

GooglePrivacyDlpV2QuasiIdField
A quasi-identifier column has a custom_tag, used to know which column in the data corresponds to which column in the statistical model.

GooglePrivacyDlpV2QuasiIdentifierField
A quasi-identifier column has a custom_tag, used to know which column in the data corresponds to which column in the statistical model.

GooglePrivacyDlpV2QuoteInfo
Message for infoType-dependent details parsed from quote.

GooglePrivacyDlpV2Range
Generic half-open interval [start, end)

GooglePrivacyDlpV2RecordCondition
A condition for determining whether a transformation should be applied to a field.

GooglePrivacyDlpV2RecordKey
Message for a unique key indicating a record that contains a finding.

GooglePrivacyDlpV2RecordLocation
Location of a finding within a row or record.

GooglePrivacyDlpV2RecordSuppression
Configuration to suppress records whose suppression conditions evaluate to true.

GooglePrivacyDlpV2RecordTransformations
A type of transformation that is applied over structured data such as a table.

GooglePrivacyDlpV2RedactConfig
Redact a given value. For example, if used with an InfoTypeTransformation transforming PHONE_NUMBER, and input 'My phone number is 206-555-0123', the output would be 'My phone number is '.

GooglePrivacyDlpV2RedactImageRequest
Request to search for potentially sensitive info in an image and redact it by covering it with a colored rectangle.

GooglePrivacyDlpV2RedactImageResponse
Results of redacting an image.

GooglePrivacyDlpV2Regex
Message defining a custom regular expression.

GooglePrivacyDlpV2ReidentifyContentRequest
Request to re-identify an item.

GooglePrivacyDlpV2ReidentifyContentResponse
Results of re-identifying a item.

GooglePrivacyDlpV2ReplaceValueConfig
Replace each input value with a given Value.

GooglePrivacyDlpV2ReplaceWithInfoTypeConfig
Replace each matching finding with the name of the info_type.

GooglePrivacyDlpV2RequestedOptions
Snapshot of the inspection configuration.

GooglePrivacyDlpV2RequestedRiskAnalysisOptions
Risk analysis options.

GooglePrivacyDlpV2Result
All result fields mentioned below are updated while the job is processing.

GooglePrivacyDlpV2RiskAnalysisJobConfig
Configuration for a risk analysis job. See https://cloud.google.com/dlp/docs/concepts-risk-analysis to learn more.

GooglePrivacyDlpV2Row
Values of the row.

GooglePrivacyDlpV2SaveFindings
If set, the detailed findings will be persisted to the specified OutputStorageConfig. Only a single instance of this action can be specified. Compatible with: Inspect, Risk

GooglePrivacyDlpV2Schedule
Schedule for inspect job triggers.

GooglePrivacyDlpV2StatisticalTable
An auxiliary table containing statistical information on the relative frequency of different quasi-identifiers values. It has one or several quasi-identifiers columns, and one column that indicates the relative frequency of each quasi-identifier tuple. If a tuple is present in the data but not in the auxiliary table, the corresponding relative frequency is assumed to be zero (and thus, the tuple is highly reidentifiable).

GooglePrivacyDlpV2StorageConfig
Shared message indicating Cloud storage type.

GooglePrivacyDlpV2StorageMetadataLabel
Storage metadata label to indicate which metadata entry contains findings.

GooglePrivacyDlpV2StoredInfoType
StoredInfoType resource message that contains information about the current version and any pending updates.

GooglePrivacyDlpV2StoredInfoTypeConfig
Configuration for stored infoTypes. All fields and subfield are provided by the user. For more information, see https://cloud.google.com/dlp/docs/creating-custom-infotypes.

GooglePrivacyDlpV2StoredInfoTypeStats
Statistics for a StoredInfoType.

GooglePrivacyDlpV2StoredInfoTypeVersion
Version of a StoredInfoType, including the configuration used to build it, create timestamp, and current state.

GooglePrivacyDlpV2StoredType
A reference to a StoredInfoType to use with scanning.

GooglePrivacyDlpV2SummaryResult
A collection that informs the user the number of times a particular TransformationResultCode and error details occurred.

GooglePrivacyDlpV2SurrogateType
Message for detecting output from deidentification transformations such as CryptoReplaceFfxFpeConfig. These types of transformations are those that perform pseudonymization, thereby producing a "surrogate" as output. This should be used in conjunction with a field on the transformation such as surrogate_info_type. This CustomInfoType does not support the use of detection_rules.

GooglePrivacyDlpV2Table
Structured content to inspect. Up to 50,000 Values per request allowed. See https://cloud.google.com/dlp/docs/inspecting-text#inspecting_a_table to learn more.

GooglePrivacyDlpV2TableLocation
Location of a finding within a table.

GooglePrivacyDlpV2TableOptions
Instructions regarding the table content being inspected.

GooglePrivacyDlpV2TaggedField
A column with a semantic tag attached.

GooglePrivacyDlpV2ThrowError
Throw an error and fail the request when a transformation error occurs.

GooglePrivacyDlpV2TimePartConfig
For use with Date, Timestamp, and TimeOfDay, extract or preserve a portion of the value.

GooglePrivacyDlpV2TimeZone
Time zone of the date time object.

GooglePrivacyDlpV2TimespanConfig
Configuration of the timespan of the items to include in scanning. Currently only supported when inspecting Google Cloud Storage and BigQuery.

GooglePrivacyDlpV2TransformationErrorHandling
How to handle transformation errors during de-identification. A transformation error occurs when the requested transformation is incompatible with the data. For example, trying to de-identify an IP address using a DateShift transformation would result in a transformation error, since date info cannot be extracted from an IP address. Information about any incompatible transformations, and how they were handled, is returned in the response as part of the TransformationOverviews.

GooglePrivacyDlpV2TransformationOverview
Overview of the modifications that occurred.

GooglePrivacyDlpV2TransformationSummary
Summary of a single transformation. Only one of 'transformation', 'field_transformation', or 'record_suppress' will be set.

GooglePrivacyDlpV2TransientCryptoKey
Use this to have a random data crypto key generated. It will be discarded after the request finishes.

GooglePrivacyDlpV2Trigger
What event needs to occur for a new job to be started.

GooglePrivacyDlpV2UnwrappedCryptoKey
Using raw keys is prone to security risks due to accidentally leaking the key. Choose another type of key if possible.

GooglePrivacyDlpV2UpdateDeidentifyTemplateRequest
Request message for UpdateDeidentifyTemplate.

GooglePrivacyDlpV2UpdateInspectTemplateRequest
Request message for UpdateInspectTemplate.

GooglePrivacyDlpV2UpdateJobTriggerRequest
Request message for UpdateJobTrigger.

GooglePrivacyDlpV2UpdateStoredInfoTypeRequest
Request message for UpdateStoredInfoType.

GooglePrivacyDlpV2Value
Set of primitive values supported by the system. Note that for the purposes of inspection or transformation, the number of bytes considered to comprise a 'Value' is based on its representation as a UTF-8 encoded string. For example, if 'integer_value' is set to 123456789, the number of bytes would be counted as 9, even though an int64 only holds up to 8 bytes of data.

GooglePrivacyDlpV2ValueFrequency
A value of a field, including its frequency.

GooglePrivacyDlpV2WordList
Message defining a list of words or phrases to search for in the data.

GoogleProtobufEmpty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

GoogleRpcStatus
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

GoogleTypeDate
Represents a whole or partial calendar date, such as a birthday. The time of day and time zone are either specified elsewhere or are insignificant. The date is relative to the Gregorian Calendar. This can represent one of the following: * A full date, with non-zero year, month, and day values * A month and day value, with a zero year, such as an anniversary * A year on its own, with zero month and day values * A year and month value, with a zero day, such as a credit card expiration date Related types are google.type.TimeOfDay and google.protobuf.Timestamp.

GoogleTypeTimeOfDay
Represents a time of day. The date and time zone are either not significant or are specified elsewhere. An API may choose to allow leap seconds. Related types are google.type.Date and google.protobuf.Timestamp.

InfoTypeListCall
Returns a list of the sensitive information types that the DLP API supports. See https://cloud.google.com/dlp/docs/infotypes-reference to learn more.

InfoTypeMethods
A builder providing access to all methods supported on infoType resources. It is not used directly, but through the DLP hub.

LocationInfoTypeListCall
Returns a list of the sensitive information types that the DLP API supports. See https://cloud.google.com/dlp/docs/infotypes-reference to learn more.

LocationMethods
A builder providing access to all methods supported on location resources. It is not used directly, but through the DLP hub.

OrganizationDeidentifyTemplateCreateCall
Creates a DeidentifyTemplate for re-using frequently used configuration for de-identifying content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

OrganizationDeidentifyTemplateDeleteCall
Deletes a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

OrganizationDeidentifyTemplateGetCall
Gets a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

OrganizationDeidentifyTemplateListCall
Lists DeidentifyTemplates. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

OrganizationDeidentifyTemplatePatchCall
Updates the DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

OrganizationInspectTemplateCreateCall
Creates an InspectTemplate for re-using frequently used configuration for inspecting content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

OrganizationInspectTemplateDeleteCall
Deletes an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

OrganizationInspectTemplateGetCall
Gets an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

OrganizationInspectTemplateListCall
Lists InspectTemplates. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

OrganizationInspectTemplatePatchCall
Updates the InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

OrganizationLocationDeidentifyTemplateCreateCall
Creates a DeidentifyTemplate for re-using frequently used configuration for de-identifying content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

OrganizationLocationDeidentifyTemplateDeleteCall
Deletes a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

OrganizationLocationDeidentifyTemplateGetCall
Gets a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

OrganizationLocationDeidentifyTemplateListCall
Lists DeidentifyTemplates. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

OrganizationLocationDeidentifyTemplatePatchCall
Updates the DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

OrganizationLocationDlpJobListCall
Lists DlpJobs that match the specified filter in the request. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

OrganizationLocationInspectTemplateCreateCall
Creates an InspectTemplate for re-using frequently used configuration for inspecting content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

OrganizationLocationInspectTemplateDeleteCall
Deletes an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

OrganizationLocationInspectTemplateGetCall
Gets an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

OrganizationLocationInspectTemplateListCall
Lists InspectTemplates. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

OrganizationLocationInspectTemplatePatchCall
Updates the InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

OrganizationLocationJobTriggerCreateCall
Creates a job trigger to run DLP actions such as scanning storage for sensitive information on a set schedule. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

OrganizationLocationJobTriggerDeleteCall
Deletes a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

OrganizationLocationJobTriggerGetCall
Gets a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

OrganizationLocationJobTriggerListCall
Lists job triggers. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

OrganizationLocationJobTriggerPatchCall
Updates a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

OrganizationLocationStoredInfoTypeCreateCall
Creates a pre-built stored infoType to be used for inspection. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

OrganizationLocationStoredInfoTypeDeleteCall
Deletes a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

OrganizationLocationStoredInfoTypeGetCall
Gets a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

OrganizationLocationStoredInfoTypeListCall
Lists stored infoTypes. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

OrganizationLocationStoredInfoTypePatchCall
Updates the stored infoType by creating a new version. The existing version will continue to be used until the new version is ready. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

OrganizationMethods
A builder providing access to all methods supported on organization resources. It is not used directly, but through the DLP hub.

OrganizationStoredInfoTypeCreateCall
Creates a pre-built stored infoType to be used for inspection. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

OrganizationStoredInfoTypeDeleteCall
Deletes a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

OrganizationStoredInfoTypeGetCall
Gets a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

OrganizationStoredInfoTypeListCall
Lists stored infoTypes. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

OrganizationStoredInfoTypePatchCall
Updates the stored infoType by creating a new version. The existing version will continue to be used until the new version is ready. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

ProjectContentDeidentifyCall
De-identifies potentially sensitive info from a ContentItem. This method has limits on input size and output size. See https://cloud.google.com/dlp/docs/deidentify-sensitive-data to learn more. When no InfoTypes or CustomInfoTypes are specified in this request, the system will automatically choose what detectors to run. By default this may be all types, but may change over time as detectors are updated.

ProjectContentInspectCall
Finds potentially sensitive info in content. This method has limits on input size, processing time, and output size. When no InfoTypes or CustomInfoTypes are specified in this request, the system will automatically choose what detectors to run. By default this may be all types, but may change over time as detectors are updated. For how to guides, see https://cloud.google.com/dlp/docs/inspecting-images and https://cloud.google.com/dlp/docs/inspecting-text,

ProjectContentReidentifyCall
Re-identifies content that has been de-identified. See https://cloud.google.com/dlp/docs/pseudonymization#re-identification_in_free_text_code_example to learn more.

ProjectDeidentifyTemplateCreateCall
Creates a DeidentifyTemplate for re-using frequently used configuration for de-identifying content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

ProjectDeidentifyTemplateDeleteCall
Deletes a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

ProjectDeidentifyTemplateGetCall
Gets a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

ProjectDeidentifyTemplateListCall
Lists DeidentifyTemplates. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

ProjectDeidentifyTemplatePatchCall
Updates the DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

ProjectDlpJobCancelCall
Starts asynchronous cancellation on a long-running DlpJob. The server makes a best effort to cancel the DlpJob, but success is not guaranteed. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

ProjectDlpJobCreateCall
Creates a new job to inspect storage or calculate risk metrics. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more. When no InfoTypes or CustomInfoTypes are specified in inspect jobs, the system will automatically choose what detectors to run. By default this may be all types, but may change over time as detectors are updated.

ProjectDlpJobDeleteCall
Deletes a long-running DlpJob. This method indicates that the client is no longer interested in the DlpJob result. The job will be cancelled if possible. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

ProjectDlpJobGetCall
Gets the latest state of a long-running DlpJob. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

ProjectDlpJobListCall
Lists DlpJobs that match the specified filter in the request. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

ProjectImageRedactCall
Redacts potentially sensitive info from an image. This method has limits on input size, processing time, and output size. See https://cloud.google.com/dlp/docs/redacting-sensitive-data-images to learn more. When no InfoTypes or CustomInfoTypes are specified in this request, the system will automatically choose what detectors to run. By default this may be all types, but may change over time as detectors are updated.

ProjectInspectTemplateCreateCall
Creates an InspectTemplate for re-using frequently used configuration for inspecting content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

ProjectInspectTemplateDeleteCall
Deletes an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

ProjectInspectTemplateGetCall
Gets an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

ProjectInspectTemplateListCall
Lists InspectTemplates. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

ProjectInspectTemplatePatchCall
Updates the InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

ProjectJobTriggerActivateCall
Activate a job trigger. Causes the immediate execute of a trigger instead of waiting on the trigger event to occur.

ProjectJobTriggerCreateCall
Creates a job trigger to run DLP actions such as scanning storage for sensitive information on a set schedule. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

ProjectJobTriggerDeleteCall
Deletes a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

ProjectJobTriggerGetCall
Gets a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

ProjectJobTriggerListCall
Lists job triggers. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

ProjectJobTriggerPatchCall
Updates a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

ProjectLocationContentDeidentifyCall
De-identifies potentially sensitive info from a ContentItem. This method has limits on input size and output size. See https://cloud.google.com/dlp/docs/deidentify-sensitive-data to learn more. When no InfoTypes or CustomInfoTypes are specified in this request, the system will automatically choose what detectors to run. By default this may be all types, but may change over time as detectors are updated.

ProjectLocationContentInspectCall
Finds potentially sensitive info in content. This method has limits on input size, processing time, and output size. When no InfoTypes or CustomInfoTypes are specified in this request, the system will automatically choose what detectors to run. By default this may be all types, but may change over time as detectors are updated. For how to guides, see https://cloud.google.com/dlp/docs/inspecting-images and https://cloud.google.com/dlp/docs/inspecting-text,

ProjectLocationContentReidentifyCall
Re-identifies content that has been de-identified. See https://cloud.google.com/dlp/docs/pseudonymization#re-identification_in_free_text_code_example to learn more.

ProjectLocationDeidentifyTemplateCreateCall
Creates a DeidentifyTemplate for re-using frequently used configuration for de-identifying content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

ProjectLocationDeidentifyTemplateDeleteCall
Deletes a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

ProjectLocationDeidentifyTemplateGetCall
Gets a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

ProjectLocationDeidentifyTemplateListCall
Lists DeidentifyTemplates. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

ProjectLocationDeidentifyTemplatePatchCall
Updates the DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

ProjectLocationDlpJobCancelCall
Starts asynchronous cancellation on a long-running DlpJob. The server makes a best effort to cancel the DlpJob, but success is not guaranteed. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

ProjectLocationDlpJobCreateCall
Creates a new job to inspect storage or calculate risk metrics. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more. When no InfoTypes or CustomInfoTypes are specified in inspect jobs, the system will automatically choose what detectors to run. By default this may be all types, but may change over time as detectors are updated.

ProjectLocationDlpJobDeleteCall
Deletes a long-running DlpJob. This method indicates that the client is no longer interested in the DlpJob result. The job will be cancelled if possible. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

ProjectLocationDlpJobFinishCall
Finish a running hybrid DlpJob. Triggers the finalization steps and running of any enabled actions that have not yet run.

ProjectLocationDlpJobGetCall
Gets the latest state of a long-running DlpJob. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

ProjectLocationDlpJobHybridInspectCall
Inspect hybrid content and store findings to a job. To review the findings, inspect the job. Inspection will occur asynchronously.

ProjectLocationDlpJobListCall
Lists DlpJobs that match the specified filter in the request. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

ProjectLocationImageRedactCall
Redacts potentially sensitive info from an image. This method has limits on input size, processing time, and output size. See https://cloud.google.com/dlp/docs/redacting-sensitive-data-images to learn more. When no InfoTypes or CustomInfoTypes are specified in this request, the system will automatically choose what detectors to run. By default this may be all types, but may change over time as detectors are updated.

ProjectLocationInspectTemplateCreateCall
Creates an InspectTemplate for re-using frequently used configuration for inspecting content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

ProjectLocationInspectTemplateDeleteCall
Deletes an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

ProjectLocationInspectTemplateGetCall
Gets an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

ProjectLocationInspectTemplateListCall
Lists InspectTemplates. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

ProjectLocationInspectTemplatePatchCall
Updates the InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

ProjectLocationJobTriggerActivateCall
Activate a job trigger. Causes the immediate execute of a trigger instead of waiting on the trigger event to occur.

ProjectLocationJobTriggerCreateCall
Creates a job trigger to run DLP actions such as scanning storage for sensitive information on a set schedule. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

ProjectLocationJobTriggerDeleteCall
Deletes a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

ProjectLocationJobTriggerGetCall
Gets a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

ProjectLocationJobTriggerHybridInspectCall
Inspect hybrid content and store findings to a trigger. The inspection will be processed asynchronously. To review the findings monitor the jobs within the trigger.

ProjectLocationJobTriggerListCall
Lists job triggers. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

ProjectLocationJobTriggerPatchCall
Updates a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

ProjectLocationStoredInfoTypeCreateCall
Creates a pre-built stored infoType to be used for inspection. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

ProjectLocationStoredInfoTypeDeleteCall
Deletes a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

ProjectLocationStoredInfoTypeGetCall
Gets a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

ProjectLocationStoredInfoTypeListCall
Lists stored infoTypes. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

ProjectLocationStoredInfoTypePatchCall
Updates the stored infoType by creating a new version. The existing version will continue to be used until the new version is ready. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the DLP hub.

ProjectStoredInfoTypeCreateCall
Creates a pre-built stored infoType to be used for inspection. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

ProjectStoredInfoTypeDeleteCall
Deletes a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

ProjectStoredInfoTypeGetCall
Gets a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

ProjectStoredInfoTypeListCall
Lists stored infoTypes. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

ProjectStoredInfoTypePatchCall
Updates the stored infoType by creating a new version. The existing version will continue to be used until the new version is ready. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.