Struct google_dlp2::api::OrganizationMethods

pub struct OrganizationMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on organization resources. It is not used directly, but through the DLP hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_dlp2 as dlp2;
 
use std::default::Default;
use oauth2;
use dlp2::DLP;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = DLP::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `deidentify_templates_create(...)`, `deidentify_templates_delete(...)`, `deidentify_templates_get(...)`, `deidentify_templates_list(...)`, `deidentify_templates_patch(...)`, `inspect_templates_create(...)`, `inspect_templates_delete(...)`, `inspect_templates_get(...)`, `inspect_templates_list(...)`, `inspect_templates_patch(...)`, `locations_deidentify_templates_create(...)`, `locations_deidentify_templates_delete(...)`, `locations_deidentify_templates_get(...)`, `locations_deidentify_templates_list(...)`, `locations_deidentify_templates_patch(...)`, `locations_dlp_jobs_list(...)`, `locations_inspect_templates_create(...)`, `locations_inspect_templates_delete(...)`, `locations_inspect_templates_get(...)`, `locations_inspect_templates_list(...)`, `locations_inspect_templates_patch(...)`, `locations_job_triggers_create(...)`, `locations_job_triggers_delete(...)`, `locations_job_triggers_get(...)`, `locations_job_triggers_list(...)`, `locations_job_triggers_patch(...)`, `locations_stored_info_types_create(...)`, `locations_stored_info_types_delete(...)`, `locations_stored_info_types_get(...)`, `locations_stored_info_types_list(...)`, `locations_stored_info_types_patch(...)`, `stored_info_types_create(...)`, `stored_info_types_delete(...)`, `stored_info_types_get(...)`, `stored_info_types_list(...)` and `stored_info_types_patch(...)`
// to build up your call.
let rb = hub.organizations();

Implementations

impl<'a, C> OrganizationMethods<'a, C>

pub fn deidentify_templates_create(
&self,
request: GooglePrivacyDlpV2CreateDeidentifyTemplateRequest,
parent: &str
) -> OrganizationDeidentifyTemplateCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a DeidentifyTemplate for re-using frequently used configuration for de-identifying content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

Arguments

request - No description provided.

parent - Required. Parent resource name. The format of this value varies depending on the scope of the request (project or organization) and whether you have specified a processing location: + Projects scope, location specified: projects/PROJECT_ID/locations/LOCATION_ID + Projects scope, no location specified (defaults to global): projects/PROJECT_ID + Organizations scope, location specified: organizations/ORG_ID/locations/LOCATION_ID + Organizations scope, no location specified (defaults to global): organizations/ORG_ID The following example parent string specifies a parent project with the identifier example-project, and specifies the europe-west3 location for processing data: parent=projects/example-project/locations/europe-west3

pub fn deidentify_templates_delete(
&self,
name: &str
) -> OrganizationDeidentifyTemplateDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

Arguments

name - Required. Resource name of the organization and deidentify template to be deleted, for example organizations/433245324/deidentifyTemplates/432452342 or projects/project-id/deidentifyTemplates/432452342.

pub fn deidentify_templates_get(
&self,
name: &str
) -> OrganizationDeidentifyTemplateGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

Arguments

name - Required. Resource name of the organization and deidentify template to be read, for example organizations/433245324/deidentifyTemplates/432452342 or projects/project-id/deidentifyTemplates/432452342.

pub fn deidentify_templates_list(
&self,
parent: &str
) -> OrganizationDeidentifyTemplateListCall<'a, C>

Create a builder to help you perform the following task:

Lists DeidentifyTemplates. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

Arguments

parent - Required. Parent resource name. The format of this value varies depending on the scope of the request (project or organization) and whether you have specified a processing location: + Projects scope, location specified: projects/PROJECT_ID/locations/LOCATION_ID + Projects scope, no location specified (defaults to global): projects/PROJECT_ID + Organizations scope, location specified: organizations/ORG_ID/locations/LOCATION_ID + Organizations scope, no location specified (defaults to global): organizations/ORG_ID The following example parent string specifies a parent project with the identifier example-project, and specifies the europe-west3 location for processing data: parent=projects/example-project/locations/europe-west3

pub fn deidentify_templates_patch(
&self,
request: GooglePrivacyDlpV2UpdateDeidentifyTemplateRequest,
name: &str
) -> OrganizationDeidentifyTemplatePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

Arguments

request - No description provided.

name - Required. Resource name of organization and deidentify template to be updated, for example organizations/433245324/deidentifyTemplates/432452342 or projects/project-id/deidentifyTemplates/432452342.

pub fn inspect_templates_create(
&self,
request: GooglePrivacyDlpV2CreateInspectTemplateRequest,
parent: &str
) -> OrganizationInspectTemplateCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates an InspectTemplate for re-using frequently used configuration for inspecting content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

Arguments

request - No description provided.

parent - Required. Parent resource name. The format of this value varies depending on the scope of the request (project or organization) and whether you have specified a processing location: + Projects scope, location specified: projects/PROJECT_ID/locations/LOCATION_ID + Projects scope, no location specified (defaults to global): projects/PROJECT_ID + Organizations scope, location specified: organizations/ORG_ID/locations/LOCATION_ID + Organizations scope, no location specified (defaults to global): organizations/ORG_ID The following example parent string specifies a parent project with the identifier example-project, and specifies the europe-west3 location for processing data: parent=projects/example-project/locations/europe-west3

pub fn inspect_templates_delete(
&self,
name: &str
) -> OrganizationInspectTemplateDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

Arguments

name - Required. Resource name of the organization and inspectTemplate to be deleted, for example organizations/433245324/inspectTemplates/432452342 or projects/project-id/inspectTemplates/432452342.

pub fn inspect_templates_get(
&self,
name: &str
) -> OrganizationInspectTemplateGetCall<'a, C>

Create a builder to help you perform the following task:

Gets an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

Arguments

name - Required. Resource name of the organization and inspectTemplate to be read, for example organizations/433245324/inspectTemplates/432452342 or projects/project-id/inspectTemplates/432452342.

pub fn inspect_templates_list(
&self,
parent: &str
) -> OrganizationInspectTemplateListCall<'a, C>

Create a builder to help you perform the following task:

Lists InspectTemplates. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

Arguments

parent - Required. Parent resource name. The format of this value varies depending on the scope of the request (project or organization) and whether you have specified a processing location: + Projects scope, location specified: projects/PROJECT_ID/locations/LOCATION_ID + Projects scope, no location specified (defaults to global): projects/PROJECT_ID + Organizations scope, location specified: organizations/ORG_ID/locations/LOCATION_ID + Organizations scope, no location specified (defaults to global): organizations/ORG_ID The following example parent string specifies a parent project with the identifier example-project, and specifies the europe-west3 location for processing data: parent=projects/example-project/locations/europe-west3

pub fn inspect_templates_patch(
&self,
request: GooglePrivacyDlpV2UpdateInspectTemplateRequest,
name: &str
) -> OrganizationInspectTemplatePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

Arguments

request - No description provided.

name - Required. Resource name of organization and inspectTemplate to be updated, for example organizations/433245324/inspectTemplates/432452342 or projects/project-id/inspectTemplates/432452342.

pub fn locations_deidentify_templates_create(
&self,
request: GooglePrivacyDlpV2CreateDeidentifyTemplateRequest,
parent: &str
) -> OrganizationLocationDeidentifyTemplateCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a DeidentifyTemplate for re-using frequently used configuration for de-identifying content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

Arguments

request - No description provided.

parent - Required. Parent resource name. The format of this value varies depending on the scope of the request (project or organization) and whether you have specified a processing location: + Projects scope, location specified: projects/PROJECT_ID/locations/LOCATION_ID + Projects scope, no location specified (defaults to global): projects/PROJECT_ID + Organizations scope, location specified: organizations/ORG_ID/locations/LOCATION_ID + Organizations scope, no location specified (defaults to global): organizations/ORG_ID The following example parent string specifies a parent project with the identifier example-project, and specifies the europe-west3 location for processing data: parent=projects/example-project/locations/europe-west3

pub fn locations_deidentify_templates_delete(
&self,
name: &str
) -> OrganizationLocationDeidentifyTemplateDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

Arguments

name - Required. Resource name of the organization and deidentify template to be deleted, for example organizations/433245324/deidentifyTemplates/432452342 or projects/project-id/deidentifyTemplates/432452342.

pub fn locations_deidentify_templates_get(
&self,
name: &str
) -> OrganizationLocationDeidentifyTemplateGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

Arguments

name - Required. Resource name of the organization and deidentify template to be read, for example organizations/433245324/deidentifyTemplates/432452342 or projects/project-id/deidentifyTemplates/432452342.

pub fn locations_deidentify_templates_list(
&self,
parent: &str
) -> OrganizationLocationDeidentifyTemplateListCall<'a, C>

Create a builder to help you perform the following task:

Lists DeidentifyTemplates. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

Arguments

parent - Required. Parent resource name. The format of this value varies depending on the scope of the request (project or organization) and whether you have specified a processing location: + Projects scope, location specified: projects/PROJECT_ID/locations/LOCATION_ID + Projects scope, no location specified (defaults to global): projects/PROJECT_ID + Organizations scope, location specified: organizations/ORG_ID/locations/LOCATION_ID + Organizations scope, no location specified (defaults to global): organizations/ORG_ID The following example parent string specifies a parent project with the identifier example-project, and specifies the europe-west3 location for processing data: parent=projects/example-project/locations/europe-west3

pub fn locations_deidentify_templates_patch(
&self,
request: GooglePrivacyDlpV2UpdateDeidentifyTemplateRequest,
name: &str
) -> OrganizationLocationDeidentifyTemplatePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

Arguments

request - No description provided.

name - Required. Resource name of organization and deidentify template to be updated, for example organizations/433245324/deidentifyTemplates/432452342 or projects/project-id/deidentifyTemplates/432452342.

pub fn locations_dlp_jobs_list(
&self,
parent: &str
) -> OrganizationLocationDlpJobListCall<'a, C>

Create a builder to help you perform the following task:

Lists DlpJobs that match the specified filter in the request. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

Arguments

parent - Required. Parent resource name. The format of this value varies depending on whether you have specified a processing location: + Projects scope, location specified: projects/PROJECT_ID/locations/LOCATION_ID + Projects scope, no location specified (defaults to global): projects/PROJECT_ID The following example parent string specifies a parent project with the identifier example-project, and specifies the europe-west3 location for processing data: parent=projects/example-project/locations/europe-west3

pub fn locations_inspect_templates_create(
&self,
request: GooglePrivacyDlpV2CreateInspectTemplateRequest,
parent: &str
) -> OrganizationLocationInspectTemplateCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates an InspectTemplate for re-using frequently used configuration for inspecting content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

Arguments

request - No description provided.

parent - Required. Parent resource name. The format of this value varies depending on the scope of the request (project or organization) and whether you have specified a processing location: + Projects scope, location specified: projects/PROJECT_ID/locations/LOCATION_ID + Projects scope, no location specified (defaults to global): projects/PROJECT_ID + Organizations scope, location specified: organizations/ORG_ID/locations/LOCATION_ID + Organizations scope, no location specified (defaults to global): organizations/ORG_ID The following example parent string specifies a parent project with the identifier example-project, and specifies the europe-west3 location for processing data: parent=projects/example-project/locations/europe-west3

pub fn locations_inspect_templates_delete(
&self,
name: &str
) -> OrganizationLocationInspectTemplateDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

Arguments

name - Required. Resource name of the organization and inspectTemplate to be deleted, for example organizations/433245324/inspectTemplates/432452342 or projects/project-id/inspectTemplates/432452342.

pub fn locations_inspect_templates_get(
&self,
name: &str
) -> OrganizationLocationInspectTemplateGetCall<'a, C>

Create a builder to help you perform the following task:

Gets an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

Arguments

name - Required. Resource name of the organization and inspectTemplate to be read, for example organizations/433245324/inspectTemplates/432452342 or projects/project-id/inspectTemplates/432452342.

pub fn locations_inspect_templates_list(
&self,
parent: &str
) -> OrganizationLocationInspectTemplateListCall<'a, C>

Create a builder to help you perform the following task:

Lists InspectTemplates. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

Arguments

parent - Required. Parent resource name. The format of this value varies depending on the scope of the request (project or organization) and whether you have specified a processing location: + Projects scope, location specified: projects/PROJECT_ID/locations/LOCATION_ID + Projects scope, no location specified (defaults to global): projects/PROJECT_ID + Organizations scope, location specified: organizations/ORG_ID/locations/LOCATION_ID + Organizations scope, no location specified (defaults to global): organizations/ORG_ID The following example parent string specifies a parent project with the identifier example-project, and specifies the europe-west3 location for processing data: parent=projects/example-project/locations/europe-west3

pub fn locations_inspect_templates_patch(
&self,
request: GooglePrivacyDlpV2UpdateInspectTemplateRequest,
name: &str
) -> OrganizationLocationInspectTemplatePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

Arguments

request - No description provided.

name - Required. Resource name of organization and inspectTemplate to be updated, for example organizations/433245324/inspectTemplates/432452342 or projects/project-id/inspectTemplates/432452342.

pub fn locations_job_triggers_create(
&self,
request: GooglePrivacyDlpV2CreateJobTriggerRequest,
parent: &str
) -> OrganizationLocationJobTriggerCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a job trigger to run DLP actions such as scanning storage for sensitive information on a set schedule. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

Arguments

request - No description provided.

parent - Required. Parent resource name. The format of this value varies depending on whether you have specified a processing location: + Projects scope, location specified: projects/PROJECT_ID/locations/LOCATION_ID + Projects scope, no location specified (defaults to global): projects/PROJECT_ID The following example parent string specifies a parent project with the identifier example-project, and specifies the europe-west3 location for processing data: parent=projects/example-project/locations/europe-west3

pub fn locations_job_triggers_delete(
&self,
name: &str
) -> OrganizationLocationJobTriggerDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

Arguments

name - Required. Resource name of the project and the triggeredJob, for example projects/dlp-test-project/jobTriggers/53234423.

pub fn locations_job_triggers_get(
&self,
name: &str
) -> OrganizationLocationJobTriggerGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

Arguments

name - Required. Resource name of the project and the triggeredJob, for example projects/dlp-test-project/jobTriggers/53234423.

pub fn locations_job_triggers_list(
&self,
parent: &str
) -> OrganizationLocationJobTriggerListCall<'a, C>

Create a builder to help you perform the following task:

Lists job triggers. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

Arguments

parent - Required. Parent resource name. The format of this value varies depending on whether you have specified a processing location: + Projects scope, location specified: projects/PROJECT_ID/locations/LOCATION_ID + Projects scope, no location specified (defaults to global): projects/PROJECT_ID The following example parent string specifies a parent project with the identifier example-project, and specifies the europe-west3 location for processing data: parent=projects/example-project/locations/europe-west3

pub fn locations_job_triggers_patch(
&self,
request: GooglePrivacyDlpV2UpdateJobTriggerRequest,
name: &str
) -> OrganizationLocationJobTriggerPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

Arguments

request - No description provided.

name - Required. Resource name of the project and the triggeredJob, for example projects/dlp-test-project/jobTriggers/53234423.

pub fn locations_stored_info_types_create(
&self,
request: GooglePrivacyDlpV2CreateStoredInfoTypeRequest,
parent: &str
) -> OrganizationLocationStoredInfoTypeCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a pre-built stored infoType to be used for inspection. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

Arguments

request - No description provided.

parent - Required. Parent resource name. The format of this value varies depending on the scope of the request (project or organization) and whether you have specified a processing location: + Projects scope, location specified: projects/PROJECT_ID/locations/LOCATION_ID + Projects scope, no location specified (defaults to global): projects/PROJECT_ID + Organizations scope, location specified: organizations/ORG_ID/locations/LOCATION_ID + Organizations scope, no location specified (defaults to global): organizations/ORG_ID The following example parent string specifies a parent project with the identifier example-project, and specifies the europe-west3 location for processing data: parent=projects/example-project/locations/europe-west3

pub fn locations_stored_info_types_delete(
&self,
name: &str
) -> OrganizationLocationStoredInfoTypeDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

Arguments

name - Required. Resource name of the organization and storedInfoType to be deleted, for example organizations/433245324/storedInfoTypes/432452342 or projects/project-id/storedInfoTypes/432452342.

pub fn locations_stored_info_types_get(
&self,
name: &str
) -> OrganizationLocationStoredInfoTypeGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

Arguments

name - Required. Resource name of the organization and storedInfoType to be read, for example organizations/433245324/storedInfoTypes/432452342 or projects/project-id/storedInfoTypes/432452342.

pub fn locations_stored_info_types_list(
&self,
parent: &str
) -> OrganizationLocationStoredInfoTypeListCall<'a, C>

Create a builder to help you perform the following task:

Lists stored infoTypes. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

Arguments

parent - Required. Parent resource name. The format of this value varies depending on the scope of the request (project or organization) and whether you have specified a processing location: + Projects scope, location specified: projects/PROJECT_ID/locations/LOCATION_ID + Projects scope, no location specified (defaults to global): projects/PROJECT_ID + Organizations scope, location specified: organizations/ORG_ID/locations/LOCATION_ID + Organizations scope, no location specified (defaults to global): organizations/ORG_ID The following example parent string specifies a parent project with the identifier example-project, and specifies the europe-west3 location for processing data: parent=projects/example-project/locations/europe-west3

pub fn locations_stored_info_types_patch(
&self,
request: GooglePrivacyDlpV2UpdateStoredInfoTypeRequest,
name: &str
) -> OrganizationLocationStoredInfoTypePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the stored infoType by creating a new version. The existing version will continue to be used until the new version is ready. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

Arguments

request - No description provided.

name - Required. Resource name of organization and storedInfoType to be updated, for example organizations/433245324/storedInfoTypes/432452342 or projects/project-id/storedInfoTypes/432452342.

pub fn stored_info_types_create(
&self,
request: GooglePrivacyDlpV2CreateStoredInfoTypeRequest,
parent: &str
) -> OrganizationStoredInfoTypeCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a pre-built stored infoType to be used for inspection. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

Arguments

request - No description provided.

parent - Required. Parent resource name. The format of this value varies depending on the scope of the request (project or organization) and whether you have specified a processing location: + Projects scope, location specified: projects/PROJECT_ID/locations/LOCATION_ID + Projects scope, no location specified (defaults to global): projects/PROJECT_ID + Organizations scope, location specified: organizations/ORG_ID/locations/LOCATION_ID + Organizations scope, no location specified (defaults to global): organizations/ORG_ID The following example parent string specifies a parent project with the identifier example-project, and specifies the europe-west3 location for processing data: parent=projects/example-project/locations/europe-west3

pub fn stored_info_types_delete(
&self,
name: &str
) -> OrganizationStoredInfoTypeDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

Arguments

name - Required. Resource name of the organization and storedInfoType to be deleted, for example organizations/433245324/storedInfoTypes/432452342 or projects/project-id/storedInfoTypes/432452342.

pub fn stored_info_types_get(
&self,
name: &str
) -> OrganizationStoredInfoTypeGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

Arguments

name - Required. Resource name of the organization and storedInfoType to be read, for example organizations/433245324/storedInfoTypes/432452342 or projects/project-id/storedInfoTypes/432452342.

pub fn stored_info_types_list(
&self,
parent: &str
) -> OrganizationStoredInfoTypeListCall<'a, C>

Create a builder to help you perform the following task:

Lists stored infoTypes. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

Arguments

parent - Required. Parent resource name. The format of this value varies depending on the scope of the request (project or organization) and whether you have specified a processing location: + Projects scope, location specified: projects/PROJECT_ID/locations/LOCATION_ID + Projects scope, no location specified (defaults to global): projects/PROJECT_ID + Organizations scope, location specified: organizations/ORG_ID/locations/LOCATION_ID + Organizations scope, no location specified (defaults to global): organizations/ORG_ID The following example parent string specifies a parent project with the identifier example-project, and specifies the europe-west3 location for processing data: parent=projects/example-project/locations/europe-west3

pub fn stored_info_types_patch(
&self,
request: GooglePrivacyDlpV2UpdateStoredInfoTypeRequest,
name: &str
) -> OrganizationStoredInfoTypePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the stored infoType by creating a new version. The existing version will continue to be used until the new version is ready. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

Arguments

request - No description provided.

name - Required. Resource name of organization and storedInfoType to be updated, for example organizations/433245324/storedInfoTypes/432452342 or projects/project-id/storedInfoTypes/432452342.