Struct google_dlp2::ProjectMethods

pub struct ProjectMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the DLP hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_dlp2 as dlp2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use dlp2::DLP;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = DLP::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `content_deidentify(...)`, `content_inspect(...)`, `content_reidentify(...)`, `deidentify_templates_create(...)`, `deidentify_templates_delete(...)`, `deidentify_templates_get(...)`, `deidentify_templates_list(...)`, `deidentify_templates_patch(...)`, `dlp_jobs_cancel(...)`, `dlp_jobs_create(...)`, `dlp_jobs_delete(...)`, `dlp_jobs_get(...)`, `dlp_jobs_list(...)`, `image_redact(...)`, `inspect_templates_create(...)`, `inspect_templates_delete(...)`, `inspect_templates_get(...)`, `inspect_templates_list(...)`, `inspect_templates_patch(...)`, `job_triggers_activate(...)`, `job_triggers_create(...)`, `job_triggers_delete(...)`, `job_triggers_get(...)`, `job_triggers_list(...)`, `job_triggers_patch(...)`, `locations_content_deidentify(...)`, `stored_info_types_create(...)`, `stored_info_types_delete(...)`, `stored_info_types_get(...)`, `stored_info_types_list(...)` and `stored_info_types_patch(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

pub fn inspect_templates_get(
&self,
name: &str
) -> ProjectInspectTemplateGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

Arguments

name - Resource name of the organization and inspectTemplate to be read, for example organizations/433245324/inspectTemplates/432452342 or projects/project-id/inspectTemplates/432452342.

pub fn deidentify_templates_get(
&self,
name: &str
) -> ProjectDeidentifyTemplateGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

Arguments

name - Resource name of the organization and deidentify template to be read, for example organizations/433245324/deidentifyTemplates/432452342 or projects/project-id/deidentifyTemplates/432452342.

pub fn inspect_templates_patch(
&self,
request: GooglePrivacyDlpV2UpdateInspectTemplateRequest,
name: &str
) -> ProjectInspectTemplatePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

Arguments

request - No description provided.

name - Resource name of organization and inspectTemplate to be updated, for example organizations/433245324/inspectTemplates/432452342 or projects/project-id/inspectTemplates/432452342.

pub fn stored_info_types_delete(
&self,
name: &str
) -> ProjectStoredInfoTypeDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

Arguments

name - Resource name of the organization and storedInfoType to be deleted, for example organizations/433245324/storedInfoTypes/432452342 or projects/project-id/storedInfoTypes/432452342.

pub fn job_triggers_patch(
&self,
request: GooglePrivacyDlpV2UpdateJobTriggerRequest,
name: &str
) -> ProjectJobTriggerPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

Arguments

request - No description provided.

name - Resource name of the project and the triggeredJob, for example projects/dlp-test-project/jobTriggers/53234423.

pub fn locations_content_deidentify(
&self,
request: GooglePrivacyDlpV2DeidentifyContentRequest,
parent: &str,
location: &str
) -> ProjectLocationContentDeidentifyCall<'a, C, A>

Create a builder to help you perform the following task:

De-identifies potentially sensitive info from a ContentItem. This method has limits on input size and output size. See https://cloud.google.com/dlp/docs/deidentify-sensitive-data to learn more.

When no InfoTypes or CustomInfoTypes are specified in this request, the system will automatically choose what detectors to run. By default this may be all types, but may change over time as detectors are updated.

Arguments

request - No description provided.

parent - The parent resource name, for example projects/my-project-id.

location - The geographic location to process de-identification. Reserved for future extensions.

pub fn content_inspect(
&self,
request: GooglePrivacyDlpV2InspectContentRequest,
parent: &str
) -> ProjectContentInspectCall<'a, C, A>

Create a builder to help you perform the following task:

Finds potentially sensitive info in content. This method has limits on input size, processing time, and output size.

When no InfoTypes or CustomInfoTypes are specified in this request, the system will automatically choose what detectors to run. By default this may be all types, but may change over time as detectors are updated.

For how to guides, see https://cloud.google.com/dlp/docs/inspecting-images and https://cloud.google.com/dlp/docs/inspecting-text,

Arguments

request - No description provided.

parent - The parent resource name, for example projects/my-project-id.

pub fn inspect_templates_delete(
&self,
name: &str
) -> ProjectInspectTemplateDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

Arguments

name - Resource name of the organization and inspectTemplate to be deleted, for example organizations/433245324/inspectTemplates/432452342 or projects/project-id/inspectTemplates/432452342.

pub fn deidentify_templates_patch(
&self,
request: GooglePrivacyDlpV2UpdateDeidentifyTemplateRequest,
name: &str
) -> ProjectDeidentifyTemplatePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

Arguments

request - No description provided.

name - Resource name of organization and deidentify template to be updated, for example organizations/433245324/deidentifyTemplates/432452342 or projects/project-id/deidentifyTemplates/432452342.

pub fn job_triggers_get(&self, name: &str) -> ProjectJobTriggerGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

Arguments

name - Resource name of the project and the triggeredJob, for example projects/dlp-test-project/jobTriggers/53234423.

pub fn stored_info_types_get(
&self,
name: &str
) -> ProjectStoredInfoTypeGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

Arguments

name - Resource name of the organization and storedInfoType to be read, for example organizations/433245324/storedInfoTypes/432452342 or projects/project-id/storedInfoTypes/432452342.

pub fn inspect_templates_create(
&self,
request: GooglePrivacyDlpV2CreateInspectTemplateRequest,
parent: &str
) -> ProjectInspectTemplateCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates an InspectTemplate for re-using frequently used configuration for inspecting content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

Arguments

request - No description provided.

parent - The parent resource name, for example projects/my-project-id or organizations/my-org-id.

pub fn job_triggers_create(
&self,
request: GooglePrivacyDlpV2CreateJobTriggerRequest,
parent: &str
) -> ProjectJobTriggerCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a job trigger to run DLP actions such as scanning storage for sensitive information on a set schedule. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

Arguments

request - No description provided.

parent - The parent resource name, for example projects/my-project-id.

pub fn dlp_jobs_create(
&self,
request: GooglePrivacyDlpV2CreateDlpJobRequest,
parent: &str
) -> ProjectDlpJobCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new job to inspect storage or calculate risk metrics. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

When no InfoTypes or CustomInfoTypes are specified in inspect jobs, the system will automatically choose what detectors to run. By default this may be all types, but may change over time as detectors are updated.

Arguments

request - No description provided.

parent - The parent resource name, for example projects/my-project-id.

pub fn stored_info_types_patch(
&self,
request: GooglePrivacyDlpV2UpdateStoredInfoTypeRequest,
name: &str
) -> ProjectStoredInfoTypePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the stored infoType by creating a new version. The existing version will continue to be used until the new version is ready. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

Arguments

request - No description provided.

name - Resource name of organization and storedInfoType to be updated, for example organizations/433245324/storedInfoTypes/432452342 or projects/project-id/storedInfoTypes/432452342.

pub fn deidentify_templates_create(
&self,
request: GooglePrivacyDlpV2CreateDeidentifyTemplateRequest,
parent: &str
) -> ProjectDeidentifyTemplateCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a DeidentifyTemplate for re-using frequently used configuration for de-identifying content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

Arguments

request - No description provided.

parent - The parent resource name, for example projects/my-project-id or organizations/my-org-id.

pub fn dlp_jobs_list(&self, parent: &str) -> ProjectDlpJobListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists DlpJobs that match the specified filter in the request. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

Arguments

parent - The parent resource name, for example projects/my-project-id.

pub fn deidentify_templates_list(
&self,
parent: &str
) -> ProjectDeidentifyTemplateListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists DeidentifyTemplates. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

Arguments

parent - The parent resource name, for example projects/my-project-id or organizations/my-org-id.

pub fn dlp_jobs_cancel(
&self,
request: GooglePrivacyDlpV2CancelDlpJobRequest,
name: &str
) -> ProjectDlpJobCancelCall<'a, C, A>

Create a builder to help you perform the following task:

Starts asynchronous cancellation on a long-running DlpJob. The server makes a best effort to cancel the DlpJob, but success is not guaranteed. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

Arguments

request - No description provided.

name - The name of the DlpJob resource to be cancelled.

pub fn inspect_templates_list(
&self,
parent: &str
) -> ProjectInspectTemplateListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists InspectTemplates. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

Arguments

parent - The parent resource name, for example projects/my-project-id or organizations/my-org-id.

pub fn dlp_jobs_get(&self, name: &str) -> ProjectDlpJobGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the latest state of a long-running DlpJob. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

Arguments

name - The name of the DlpJob resource.

pub fn content_deidentify(
&self,
request: GooglePrivacyDlpV2DeidentifyContentRequest,
parent: &str
) -> ProjectContentDeidentifyCall<'a, C, A>

Create a builder to help you perform the following task:

De-identifies potentially sensitive info from a ContentItem. This method has limits on input size and output size. See https://cloud.google.com/dlp/docs/deidentify-sensitive-data to learn more.

When no InfoTypes or CustomInfoTypes are specified in this request, the system will automatically choose what detectors to run. By default this may be all types, but may change over time as detectors are updated.

Arguments

request - No description provided.

parent - The parent resource name, for example projects/my-project-id.

pub fn job_triggers_activate(
&self,
request: GooglePrivacyDlpV2ActivateJobTriggerRequest,
name: &str
) -> ProjectJobTriggerActivateCall<'a, C, A>

Create a builder to help you perform the following task:

Activate a job trigger. Causes the immediate execute of a trigger instead of waiting on the trigger event to occur.

Arguments

request - No description provided.

name - Resource name of the trigger to activate, for example projects/dlp-test-project/jobTriggers/53234423.

pub fn deidentify_templates_delete(
&self,
name: &str
) -> ProjectDeidentifyTemplateDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

Arguments

name - Resource name of the organization and deidentify template to be deleted, for example organizations/433245324/deidentifyTemplates/432452342 or projects/project-id/deidentifyTemplates/432452342.

pub fn stored_info_types_create(
&self,
request: GooglePrivacyDlpV2CreateStoredInfoTypeRequest,
parent: &str
) -> ProjectStoredInfoTypeCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a pre-built stored infoType to be used for inspection. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

Arguments

request - No description provided.

parent - The parent resource name, for example projects/my-project-id or organizations/my-org-id.

pub fn job_triggers_delete(
&self,
name: &str
) -> ProjectJobTriggerDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

Arguments

name - Resource name of the project and the triggeredJob, for example projects/dlp-test-project/jobTriggers/53234423.

pub fn content_reidentify(
&self,
request: GooglePrivacyDlpV2ReidentifyContentRequest,
parent: &str
) -> ProjectContentReidentifyCall<'a, C, A>

Create a builder to help you perform the following task:

Re-identifies content that has been de-identified. See https://cloud.google.com/dlp/docs/pseudonymization#re-identification_in_free_text_code_example to learn more.

Arguments

request - No description provided.

parent - The parent resource name.

pub fn job_triggers_list(
&self,
parent: &str
) -> ProjectJobTriggerListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists job triggers. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

Arguments

parent - The parent resource name, for example projects/my-project-id.

pub fn dlp_jobs_delete(&self, name: &str) -> ProjectDlpJobDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a long-running DlpJob. This method indicates that the client is no longer interested in the DlpJob result. The job will be cancelled if possible. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

Arguments

name - The name of the DlpJob resource to be deleted.

pub fn image_redact(
&self,
request: GooglePrivacyDlpV2RedactImageRequest,
parent: &str
) -> ProjectImageRedactCall<'a, C, A>

Create a builder to help you perform the following task:

Redacts potentially sensitive info from an image. This method has limits on input size, processing time, and output size. See https://cloud.google.com/dlp/docs/redacting-sensitive-data-images to learn more.

When no InfoTypes or CustomInfoTypes are specified in this request, the system will automatically choose what detectors to run. By default this may be all types, but may change over time as detectors are updated.

Arguments

request - No description provided.

parent - The parent resource name, for example projects/my-project-id.

pub fn stored_info_types_list(
&self,
parent: &str
) -> ProjectStoredInfoTypeListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists stored infoTypes. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

Arguments

parent - The parent resource name, for example projects/my-project-id or organizations/my-org-id.